Accessor functfons

This set of fucntions defines helper functions for accessing pre-defined epivars in your linelist object.

get_epivars(x, ..., simplify = TRUE)

id(x)

date_onset(x)

date_report(x)

date_outcome(x)

case_definition(x)

outcome(x)

gender(x)

age(x)

age_group(x)

geo(x)

Arguments

x
a linelist object

...
epivars to be used; can be provided as characters, integers, or logical; integers and logicals are to be matched against the vector of epi variables returned by list_epivars(x, simple = TRUE). If nothing is provided, all epivars are returned.

simplify
if TRUE (default) and a single epivar is specified, it will be converted to a vector. Otherwise, a data frame is returned.

Value

a vector if the epivar can be represented by a single column, a data frame otherwise (e.g. a location column).

Note

For the geo() accessor, which necessarily contains two columns for longitude and latitude data, when simplify = TRUE, a matrix will be returned.

Examples

## make toy data
toy_data <- messy_data()
cleaned_data <- clean_data(toy_data)
ll <- as_linelist(cleaned_data,
                  id = "id",
                  date_onset = "date_of_onset",
                  gender = "gender",
                  geo_lon = "lon",
                  geo_lat = "lat"
                 )

## general purpose accessor
list_epivars(ll, simple = TRUE)

#> [1] "id"         "date_onset" "gender"     "geo_lat"    "geo_lon"   

get_epivars(ll) # all epi variable

#>        id date_of_onset gender       lat      lon
#> 1  lvijbf    2018-01-08 female 11.582811 48.26727
#> 2  glzvne    2018-01-06 female 12.582206 46.19974
#> 3  gslqeq    2018-01-07   male 13.704558 46.68366
#> 4  mfaamr    2018-01-05   male 16.489427 47.76145
#> 5  ewkadg    2018-01-07   male  9.714188 46.85687
#> 6  wxbhis    2018-01-10 female 11.441429 47.21552
#> 7  ifmmyu    2018-01-06 female 16.078899 47.17223
#> 8  cmrwtk    2018-01-06 female 14.400111 48.94434
#> 9  dbirsl    2018-01-09   male 13.235185 47.36031
#> 10 sqjlmp    2018-01-03   male 12.432755 45.82377
#> 11 gudacg    2018-01-08   male 13.995818 47.78286
#> 12 qflgkj    2018-01-06   male 16.219886 48.57991
#> 13 hvqlto    2018-01-06   male 16.722272 49.95717
#> 14 rlvuzj    2018-01-03 female 13.772498 47.04740
#> 15 cdmfnb    2018-01-07 female 10.062713 46.13261
#> 16 yesrej    2018-01-05   male 14.435050 47.03233
#> 17 ldsxty    2018-01-03   male 13.263820 48.79245
#> 18 qwtxjx    2018-01-04   male  9.268153 49.81503
#> 19 wiyjis    2018-01-10 female 11.158673 47.06303
#> 20 ysoqow    2018-01-07 female 15.408058 48.86130

get_epivars(ll, "id", "date_onset", "gender") # named epi variables

#>        id date_of_onset gender
#> 1  lvijbf    2018-01-08 female
#> 2  glzvne    2018-01-06 female
#> 3  gslqeq    2018-01-07   male
#> 4  mfaamr    2018-01-05   male
#> 5  ewkadg    2018-01-07   male
#> 6  wxbhis    2018-01-10 female
#> 7  ifmmyu    2018-01-06 female
#> 8  cmrwtk    2018-01-06 female
#> 9  dbirsl    2018-01-09   male
#> 10 sqjlmp    2018-01-03   male
#> 11 gudacg    2018-01-08   male
#> 12 qflgkj    2018-01-06   male
#> 13 hvqlto    2018-01-06   male
#> 14 rlvuzj    2018-01-03 female
#> 15 cdmfnb    2018-01-07 female
#> 16 yesrej    2018-01-05   male
#> 17 ldsxty    2018-01-03   male
#> 18 qwtxjx    2018-01-04   male
#> 19 wiyjis    2018-01-10 female
#> 20 ysoqow    2018-01-07 female

date_onset(ll)

#>  [1] "2018-01-08" "2018-01-06" "2018-01-07" "2018-01-05" "2018-01-07"
#>  [6] "2018-01-10" "2018-01-06" "2018-01-06" "2018-01-09" "2018-01-03"
#> [11] "2018-01-08" "2018-01-06" "2018-01-06" "2018-01-03" "2018-01-07"
#> [16] "2018-01-05" "2018-01-03" "2018-01-04" "2018-01-10" "2018-01-07"

id(ll)

#>  [1] lvijbf glzvne gslqeq mfaamr ewkadg wxbhis ifmmyu cmrwtk dbirsl sqjlmp
#> [11] gudacg qflgkj hvqlto rlvuzj cdmfnb yesrej ldsxty qwtxjx wiyjis ysoqow
#> 20 Levels: cdmfnb cmrwtk dbirsl ewkadg glzvne gslqeq gudacg hvqlto ... ysoqow

gender(ll)

#>  [1] female female male   male   male   female female female male   male  
#> [11] male   male   male   female female male   male   male   female female
#> Levels: female male

geo(ll) # a matrix with two columns

#>         lon       lat
#> 1  48.26727 11.582811
#> 2  46.19974 12.582206
#> 3  46.68366 13.704558
#> 4  47.76145 16.489427
#> 5  46.85687  9.714188
#> 6  47.21552 11.441429
#> 7  47.17223 16.078899
#> 8  48.94434 14.400111
#> 9  47.36031 13.235185
#> 10 45.82377 12.432755
#> 11 47.78286 13.995818
#> 12 48.57991 16.219886
#> 13 49.95717 16.722272
#> 14 47.04740 13.772498
#> 15 46.13261 10.062713
#> 16 47.03233 14.435050
#> 17 48.79245 13.263820
#> 18 49.81503  9.268153
#> 19 47.06303 11.158673
#> 20 48.86130 15.408058

get_epivars(ll, "geo_lon", "geo_lat", simplify = FALSE) # a data frame with two columns

#>         lon       lat
#> 1  48.26727 11.582811
#> 2  46.19974 12.582206
#> 3  46.68366 13.704558
#> 4  47.76145 16.489427
#> 5  46.85687  9.714188
#> 6  47.21552 11.441429
#> 7  47.17223 16.078899
#> 8  48.94434 14.400111
#> 9  47.36031 13.235185
#> 10 45.82377 12.432755
#> 11 47.78286 13.995818
#> 12 48.57991 16.219886
#> 13 49.95717 16.722272
#> 14 47.04740 13.772498
#> 15 46.13261 10.062713
#> 16 47.03233 14.435050
#> 17 48.79245 13.263820
#> 18 49.81503  9.268153
#> 19 47.06303 11.158673
#> 20 48.86130 15.408058

# epivars that haven't been defined for the data set will return an error
try(date_report(ll))

#> Error in get_epivars(x, "date_report") : 
#>   The following epivars were not found in the data:
#> date_report