Add or remove elements from the current dictionary

These functions give a finer set of controls to the dictionary, allowing the user to add rows to the dictionary.

add_epivar(epivar = NULL, hxl = "", description = "")

add_description(epivar = NULL, description = NULL)

Arguments

epivar
a character describing an epivar or a data frame containing a dictionary to append to the current dictionary.

hxl
a hxl hashtag and metadata

description
a character string giving a human-readable description of the epivar.

Value

NULL, invisibly

Examples

# preserve current dictionary
od <- get_dictionary()

# add a single row for "date_hospital"
hosp <- data.frame(
  epivar = "date_hospital",
  hxl    = "#date +start",
  description = "date at which patient was hospitalized",
  stringsAsFactors = FALSE
)

add_epivar(hosp)
get_dictionary()

#>             epivar              hxl
#> 1               id       #respondee
#> 2       date_onset     #date +start
#> 3      date_report  #date +reported
#> 4     date_outcome       #date +end
#> 5  case_definition #indicator +name
#> 6          outcome #indicator +type
#> 7           gender #indicator +type
#> 8              age  #indicator +num
#> 9        age_group #indicator +type
#> 10         geo_lon        #geo +lon
#> 11         geo_lat        #geo +lat
#> 12   date_hospital     #date +start
#>                                            description
#> 1                         unique individual identifier
#> 2                       date at which symptoms started
#> 3                      date at which case was reported
#> 4              date of the outcome (recovery or death)
#> 5  case type: suspected, probable, confirmed, negative
#> 6                                    recovery or death
#> 7                             gender of the individual
#> 8                       age of the individual in years
#> 9                 age group of the individual in years
#> 10                    geographic coordinate: longitude
#> 11                     geographic coordinate: latitude
#> 12              date at which patient was hospitalized

# Use add_description to update descriptions
add_description("date_hospital", "hospitalisation date")
get_dictionary()

#>             epivar              hxl
#> 1               id       #respondee
#> 2       date_onset     #date +start
#> 3      date_report  #date +reported
#> 4     date_outcome       #date +end
#> 5  case_definition #indicator +name
#> 6          outcome #indicator +type
#> 7           gender #indicator +type
#> 8              age  #indicator +num
#> 9        age_group #indicator +type
#> 10         geo_lon        #geo +lon
#> 11         geo_lat        #geo +lat
#> 12   date_hospital     #date +start
#>                                            description
#> 1                         unique individual identifier
#> 2                       date at which symptoms started
#> 3                      date at which case was reported
#> 4              date of the outcome (recovery or death)
#> 5  case type: suspected, probable, confirmed, negative
#> 6                                    recovery or death
#> 7                             gender of the individual
#> 8                       age of the individual in years
#> 9                 age group of the individual in years
#> 10                    geographic coordinate: longitude
#> 11                     geographic coordinate: latitude
#> 12                                hospitalisation date

# You can export your dictionary for later use
tmp <- tempdir()
rio::export(get_dictionary(), file.path(tmp, "new_dictionary.csv"), format = "|")
head(rio::import(file.path(tmp, "new_dictionary.csv"), format = "|"))

#>            epivar              hxl
#> 1              id       #respondee
#> 2      date_onset     #date +start
#> 3     date_report  #date +reported
#> 4    date_outcome       #date +end
#> 5 case_definition #indicator +name
#> 6         outcome #indicator +type
#>                                           description
#> 1                        unique individual identifier
#> 2                      date at which symptoms started
#> 3                     date at which case was reported
#> 4             date of the outcome (recovery or death)
#> 5 case type: suspected, probable, confirmed, negative
#> 6                                   recovery or death

unlink(file.path(tmp, "new_dictionary.csv")) # cleanup

# reset dictionary
set_dictionary(od)