Clean a data.frame

This function applies several cleaning procedures to an input data.frame, by standardising variable names, labels used categorical variables (characters of factors), and setting dates to Date objects. Optionally, an intelligent date search can be used on character strings to extract dates from various formats mixed with other text. See details for more information.

clean_data(x, sep = "_", force_Date = TRUE, guess_dates = TRUE,
  error_tolerance = 0.5, quiet = FALSE)

Arguments

x
a data.frame

sep
The separator used between words, and defaults to the underscore _.

force_Date
a logical indicating if POSIXct and POSIXlt objects should be converted to Date objects; defaults to TRUE; you should use this if your dates are only precise to the day (i.e. no time information within days).

guess_dates
a logical indicating if dates should be guessed in columns storing character strings or factors; this feature is experimental; see guess_dates() for more information.

error_tolerance
a number between 0 and 1 indicating the proportion of entries which cannot be identified as dates to be tolerated; if this proportion is exceeded, the original vector is returned, and a message is issued; defaults to 0.1 (10 percent)

quiet
a logical indicating if messages should be displayed to the console (TRUE, default); set to FALSE to silence messages

Value

A data.frame with standardised labels for characters and factors.

Examples

## make toy data
onsets <- as.Date("2018-01-01") + sample(1:10, 20, replace = TRUE)
discharge <- format(as.Date(onsets) + 10, "%d/%m/%Y")
genders <- c("male", "female", "FEMALE", "Male", "Female", "MALE")
gender <- sample(genders, 20, replace = TRUE)
case_types <- c("confirmed", "probable", "suspected", "not a case",
                "Confirmed", "PROBABLE", "suspected  ", "Not.a.Case")
messy_dates <- sample(
                 c("01-12-2001", "male", "female", "2018-10-18", "2018_10_17",
                   "2018 10 19", "// 24//12//1989", NA, "that's 24/12/1989!"),
                 20, replace = TRUE)
case <- factor(sample(case_types, 20, replace = TRUE))
toy_data <- data.frame("Date of Onset." = onsets,
                       "DisCharge.." = discharge,
                       "GENDER_ " = gender,
                       "Épi.Case_définition" = case,
                       "messy/dates" = messy_dates)
## show data
toy_data

#>    Date.of.Onset. DisCharge.. GENDER_. Épi.Case_définition        messy.dates
#> 1      2018-01-04  14/01/2018   FEMALE            PROBABLE         2018 10 19
#> 2      2018-01-09  19/01/2018   female         suspected      // 24//12//1989
#> 3      2018-01-06  16/01/2018     MALE           suspected             female
#> 4      2018-01-08  18/01/2018     male           Confirmed         2018 10 19
#> 5      2018-01-09  19/01/2018   Female          not a case               male
#> 6      2018-01-02  12/01/2018   FEMALE            probable         2018_10_17
#> 7      2018-01-11  21/01/2018     MALE           Confirmed               male
#> 8      2018-01-04  14/01/2018     male            probable               male
#> 9      2018-01-11  21/01/2018   FEMALE            PROBABLE         2018-10-18
#> 10     2018-01-07  17/01/2018     Male            probable that's 24/12/1989!
#> 11     2018-01-11  21/01/2018     MALE           Confirmed               <NA>
#> 12     2018-01-06  16/01/2018     male           confirmed             female
#> 13     2018-01-05  15/01/2018     male           Confirmed         2018 10 19
#> 14     2018-01-06  16/01/2018   Female           confirmed               male
#> 15     2018-01-11  21/01/2018   female         suspected               female
#> 16     2018-01-04  14/01/2018     male           suspected         2018 10 19
#> 17     2018-01-04  14/01/2018     MALE            probable         2018_10_17
#> 18     2018-01-10  20/01/2018     male          not a case    // 24//12//1989
#> 19     2018-01-03  13/01/2018   FEMALE            PROBABLE that's 24/12/1989!
#> 20     2018-01-10  20/01/2018     MALE          Not.a.Case         01-12-2001

## clean variable names, store in new object, show results
clean_data <- clean_data(toy_data, error_tolerance = 0.1)
clean_data

#>    date_of_onset  discharge gender epi_case_definition       messy_dates
#> 1     2018-01-04 2018-01-14 female            probable        2018_10_19
#> 2     2018-01-09 2018-01-19 female           suspected        24_12_1989
#> 3     2018-01-06 2018-01-16   male           suspected            female
#> 4     2018-01-08 2018-01-18   male           confirmed        2018_10_19
#> 5     2018-01-09 2018-01-19 female          not_a_case              male
#> 6     2018-01-02 2018-01-12 female            probable        2018_10_17
#> 7     2018-01-11 2018-01-21   male           confirmed              male
#> 8     2018-01-04 2018-01-14   male            probable              male
#> 9     2018-01-11 2018-01-21 female            probable        2018_10_18
#> 10    2018-01-07 2018-01-17   male            probable that_s_24_12_1989
#> 11    2018-01-11 2018-01-21   male           confirmed              <NA>
#> 12    2018-01-06 2018-01-16   male           confirmed            female
#> 13    2018-01-05 2018-01-15   male           confirmed        2018_10_19
#> 14    2018-01-06 2018-01-16 female           confirmed              male
#> 15    2018-01-11 2018-01-21 female           suspected            female
#> 16    2018-01-04 2018-01-14   male           suspected        2018_10_19
#> 17    2018-01-04 2018-01-14   male            probable        2018_10_17
#> 18    2018-01-10 2018-01-20   male          not_a_case        24_12_1989
#> 19    2018-01-03 2018-01-13 female            probable that_s_24_12_1989
#> 20    2018-01-10 2018-01-20   male          not_a_case        01_12_2001

clean_data2 <- clean_data(toy_data, error_tolerance = 0.8)
clean_data2

#>    date_of_onset  discharge gender epi_case_definition messy_dates
#> 1     2018-01-04 2018-01-14 female            probable  2018-10-19
#> 2     2018-01-09 2018-01-19 female           suspected  1989-12-24
#> 3     2018-01-06 2018-01-16   male           suspected        <NA>
#> 4     2018-01-08 2018-01-18   male           confirmed  2018-10-19
#> 5     2018-01-09 2018-01-19 female          not_a_case        <NA>
#> 6     2018-01-02 2018-01-12 female            probable  2018-10-17
#> 7     2018-01-11 2018-01-21   male           confirmed        <NA>
#> 8     2018-01-04 2018-01-14   male            probable        <NA>
#> 9     2018-01-11 2018-01-21 female            probable  2018-10-18
#> 10    2018-01-07 2018-01-17   male            probable  1989-12-24
#> 11    2018-01-11 2018-01-21   male           confirmed        <NA>
#> 12    2018-01-06 2018-01-16   male           confirmed        <NA>
#> 13    2018-01-05 2018-01-15   male           confirmed  2018-10-19
#> 14    2018-01-06 2018-01-16 female           confirmed        <NA>
#> 15    2018-01-11 2018-01-21 female           suspected        <NA>
#> 16    2018-01-04 2018-01-14   male           suspected  2018-10-19
#> 17    2018-01-04 2018-01-14   male            probable  2018-10-17
#> 18    2018-01-10 2018-01-20   male          not_a_case  1989-12-24
#> 19    2018-01-03 2018-01-13 female            probable  1989-12-24
#> 20    2018-01-10 2018-01-20   male          not_a_case  2001-12-01