Handle dates data

This function detects variables of data.frame which are effectively representing dates, and converts them to Date objects. When variables are character strings or factors, the function will try to convert dates with various pre-defined formats (see details). For each variable, the most common date format is automatically detected, and dates not following it are set to NA (i.e. missing). It uses a tolerance threshold for the amount of entries which cannot be converted to date (error_tolerance). By default, tolerance is set to 0.1, meaning 10 for a given variable. If there are more errors, this variable is assumed not to be a date, and left untouched.

clean_dates(x, force_Date = TRUE, guess_dates = TRUE,
  error_tolerance = 0.5)

Arguments

x
a data.frame

force_Date
a logical indicating if POSIXct and POSIXlt objects should be converted to Date objects; defaults to TRUE; you should use this if your dates are only precise to the day (i.e. no time information within days).

guess_dates
a logical indicating if dates should be guessed in columns storing character strings or factors; this feature is experimental; see guess_dates() for more information.

error_tolerance
a number between 0 and 1 indicating the proportion of entries which cannot be identified as dates to be tolerated; if this proportion is exceeded, the original vector is returned, and a message is issued; defaults to 0.1 (10 percent)

Value

A data.frame with standardised dates.

See also

Examples

## make toy data
onsets <- as.POSIXct("2018-01-01") + sample(1:10, 20, replace = TRUE)
onsets2 <- format(as.Date(onsets), "%d/%m/%Y")
onsets3 <- format(as.Date(onsets), "%d %m %Y")
outcomes <- onsets + 1e7
admissions <- as.character(as.Date(onsets) + 1)
admissions[1:5] <- NA
discharges <- factor(as.Date(admissions) + 1)
onset_with_errors <- onsets2
onset_with_errors[c(1,20)] <- c("male", "confirmed")
mixed_info <- onsets3
mixed_info[1:10] <- sample(c("bleeding", "fever"), 10, replace = TRUE)
gender <- sample(c("male", "female"), 20, replace = TRUE)
case_type <- c("confirmed", "probable", "suspected", "not a case")
case <- sample(case_type, 20, replace = TRUE)
toy_data <- data.frame("Date of Onset." = onsets,
                       "onset 2" = onsets2,
                       "ONSET 3" = onsets3,
                       "onset_4" = onset_with_errors,
                       "date admission" = admissions,
                       "DATE.of.DISCHARGE" = discharges,
                       "GENDER_ " = gender,
                       "Épi.Case_définition" = case,
                       "date of admission" = admissions,
                       "Date-of_discharge" = discharges,
                       "extra" = mixed_info,
                       stringsAsFactors = FALSE)
## show data
toy_data

#>         Date.of.Onset.    onset.2    ONSET.3    onset_4 date.admission
#> 1  2018-01-01 00:00:04 01/01/2018 01 01 2018       male           <NA>
#> 2  2018-01-01 00:00:07 01/01/2018 01 01 2018 01/01/2018           <NA>
#> 3  2018-01-01 00:00:09 01/01/2018 01 01 2018 01/01/2018           <NA>
#> 4  2018-01-01 00:00:02 01/01/2018 01 01 2018 01/01/2018           <NA>
#> 5  2018-01-01 00:00:10 01/01/2018 01 01 2018 01/01/2018           <NA>
#> 6  2018-01-01 00:00:01 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 7  2018-01-01 00:00:03 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 8  2018-01-01 00:00:07 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 9  2018-01-01 00:00:06 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 10 2018-01-01 00:00:03 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 11 2018-01-01 00:00:10 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 12 2018-01-01 00:00:07 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 13 2018-01-01 00:00:01 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 14 2018-01-01 00:00:07 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 15 2018-01-01 00:00:09 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 16 2018-01-01 00:00:10 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 17 2018-01-01 00:00:05 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 18 2018-01-01 00:00:05 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 19 2018-01-01 00:00:04 01/01/2018 01 01 2018 01/01/2018     2018-01-02
#> 20 2018-01-01 00:00:03 01/01/2018 01 01 2018  confirmed     2018-01-02
#>    DATE.of.DISCHARGE GENDER_. Épi.Case_définition date.of.admission
#> 1               <NA>     male           suspected              <NA>
#> 2               <NA>     male           suspected              <NA>
#> 3               <NA>   female           suspected              <NA>
#> 4               <NA>   female           suspected              <NA>
#> 5               <NA>   female           confirmed              <NA>
#> 6         2018-01-03     male            probable        2018-01-02
#> 7         2018-01-03     male          not a case        2018-01-02
#> 8         2018-01-03     male           suspected        2018-01-02
#> 9         2018-01-03     male           suspected        2018-01-02
#> 10        2018-01-03     male          not a case        2018-01-02
#> 11        2018-01-03     male           suspected        2018-01-02
#> 12        2018-01-03     male          not a case        2018-01-02
#> 13        2018-01-03     male            probable        2018-01-02
#> 14        2018-01-03   female            probable        2018-01-02
#> 15        2018-01-03     male          not a case        2018-01-02
#> 16        2018-01-03   female           suspected        2018-01-02
#> 17        2018-01-03     male           confirmed        2018-01-02
#> 18        2018-01-03   female            probable        2018-01-02
#> 19        2018-01-03   female           confirmed        2018-01-02
#> 20        2018-01-03     male            probable        2018-01-02
#>    Date.of_discharge      extra
#> 1               <NA>      fever
#> 2               <NA>   bleeding
#> 3               <NA>   bleeding
#> 4               <NA>      fever
#> 5               <NA>   bleeding
#> 6         2018-01-03   bleeding
#> 7         2018-01-03      fever
#> 8         2018-01-03   bleeding
#> 9         2018-01-03   bleeding
#> 10        2018-01-03   bleeding
#> 11        2018-01-03 01 01 2018
#> 12        2018-01-03 01 01 2018
#> 13        2018-01-03 01 01 2018
#> 14        2018-01-03 01 01 2018
#> 15        2018-01-03 01 01 2018
#> 16        2018-01-03 01 01 2018
#> 17        2018-01-03 01 01 2018
#> 18        2018-01-03 01 01 2018
#> 19        2018-01-03 01 01 2018
#> 20        2018-01-03 01 01 2018

str(toy_data)

#> 'data.frame':	20 obs. of  11 variables:
#>  $ Date.of.Onset.     : POSIXct, format: "2018-01-01 00:00:04" "2018-01-01 00:00:07" ...
#>  $ onset.2            : chr  "01/01/2018" "01/01/2018" "01/01/2018" "01/01/2018" ...
#>  $ ONSET.3            : chr  "01 01 2018" "01 01 2018" "01 01 2018" "01 01 2018" ...
#>  $ onset_4            : chr  "male" "01/01/2018" "01/01/2018" "01/01/2018" ...
#>  $ date.admission     : chr  NA NA NA NA ...
#>  $ DATE.of.DISCHARGE  : Factor w/ 1 level "2018-01-03": NA NA NA NA NA 1 1 1 1 1 ...
#>  $ GENDER_.           : chr  "male" "male" "female" "female" ...
#>  $ Épi.Case_définition: chr  "suspected" "suspected" "suspected" "suspected" ...
#>  $ date.of.admission  : chr  NA NA NA NA ...
#>  $ Date.of_discharge  : Factor w/ 1 level "2018-01-03": NA NA NA NA NA 1 1 1 1 1 ...
#>  $ extra              : chr  "fever" "bleeding" "bleeding" "fever" ...

## clean variable names, store in new object, show results
clean_data <- clean_variable_names(toy_data)
clean_data <- clean_dates(clean_data)
clean_data

#>    date_of_onset    onset_2    onset_3    onset_4 date_admission
#> 1     2018-01-01 2018-01-01 2018-01-01       <NA>           <NA>
#> 2     2018-01-01 2018-01-01 2018-01-01 2018-01-01           <NA>
#> 3     2018-01-01 2018-01-01 2018-01-01 2018-01-01           <NA>
#> 4     2018-01-01 2018-01-01 2018-01-01 2018-01-01           <NA>
#> 5     2018-01-01 2018-01-01 2018-01-01 2018-01-01           <NA>
#> 6     2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 7     2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 8     2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 9     2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 10    2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 11    2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 12    2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 13    2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 14    2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 15    2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 16    2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 17    2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 18    2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 19    2018-01-01 2018-01-01 2018-01-01 2018-01-01     2018-01-02
#> 20    2018-01-01 2018-01-01 2018-01-01       <NA>     2018-01-02
#>    date_of_discharge gender epi_case_definition date_of_admission
#> 1               <NA>   male           suspected              <NA>
#> 2               <NA>   male           suspected              <NA>
#> 3               <NA> female           suspected              <NA>
#> 4               <NA> female           suspected              <NA>
#> 5               <NA> female           confirmed              <NA>
#> 6         2018-01-03   male            probable        2018-01-02
#> 7         2018-01-03   male          not a case        2018-01-02
#> 8         2018-01-03   male           suspected        2018-01-02
#> 9         2018-01-03   male           suspected        2018-01-02
#> 10        2018-01-03   male          not a case        2018-01-02
#> 11        2018-01-03   male           suspected        2018-01-02
#> 12        2018-01-03   male          not a case        2018-01-02
#> 13        2018-01-03   male            probable        2018-01-02
#> 14        2018-01-03 female            probable        2018-01-02
#> 15        2018-01-03   male          not a case        2018-01-02
#> 16        2018-01-03 female           suspected        2018-01-02
#> 17        2018-01-03   male           confirmed        2018-01-02
#> 18        2018-01-03 female            probable        2018-01-02
#> 19        2018-01-03 female           confirmed        2018-01-02
#> 20        2018-01-03   male            probable        2018-01-02
#>    date_of_discharge      extra
#> 1               <NA>       <NA>
#> 2               <NA>       <NA>
#> 3               <NA>       <NA>
#> 4               <NA>       <NA>
#> 5               <NA>       <NA>
#> 6         2018-01-03       <NA>
#> 7         2018-01-03       <NA>
#> 8         2018-01-03       <NA>
#> 9         2018-01-03       <NA>
#> 10        2018-01-03       <NA>
#> 11        2018-01-03 2018-01-01
#> 12        2018-01-03 2018-01-01
#> 13        2018-01-03 2018-01-01
#> 14        2018-01-03 2018-01-01
#> 15        2018-01-03 2018-01-01
#> 16        2018-01-03 2018-01-01
#> 17        2018-01-03 2018-01-01
#> 18        2018-01-03 2018-01-01
#> 19        2018-01-03 2018-01-01
#> 20        2018-01-03 2018-01-01