Check and clean spelling or codes of multiple variables in a data frame

This function allows you to clean your data according to pre-defined rules encapsulated in either a data frame or list of data frames. It has application for addressing mis-spellings and recoding variables (e.g. from electronic survey data).

clean_variable_spelling(x = data.frame(), wordlists = list(),
  spelling_vars = 3, sort_by = NULL, classes = NULL, warn = FALSE)

Arguments

x
a data.frame

wordlists
a data frame or named list of data frames with at least two columns defining the word list to be used. If this is a data frame, a third column must be present to split the wordlists by column in x (see spelling_vars).

spelling_vars
character or integer. If wordlists is a data frame, then this column in defines the columns in x corresponding to each section of the wordlists data frame. This defaults to 3, indicating the third column is to be used.

sort_by
a character the column to be used for sorting the values in each data frame. If the incoming variables are factors, this determines how the resulting factors will be sorted.

classes
a vector of class definitions for each of the columns. If this is not provided, the classes will be read from the columns themselves. Practically, this is used in clean_data() to mark columns as protected.

warn
if TRUE, warnings and errors from clean_spelling() will be shown as a single warning. Defaults to FALSE, which shows nothing.

Value

a data frame with re-defined data based on the dictionary

Details

By default, this applies the function clean_spelling() to all columns specified by the column names listed in spelling_vars, or, if a global dictionary is used, this includes all character and factor columns as well.

Global wordlists
A global wordlist is a set of definitions applied to all valid columns of x indiscriminantly.

.global spelling_var: If you want to apply a set of definitions to all valid columns in addition to specified columns, then you can include a .global group in the spelling_var column of your wordlists data frame. This is useful for setting up a dictionary of common spelling errors. NOTE: specific variable definitions will override global defintions. For example: if you have a column for cardinal directions and a definiton for N = North, then the global variable N = no will not override that. See Example.

spelling_var = NULL: If you want your data frame to be applied to all character/factor columns indiscriminantly, then setting spelling_var = NULL will use that wordlist globally.

Note

This function will only parse character and factor columns to protect numeric and Date columns from conversion to character.

See also

Examples

# Set up wordlist ------------------------------------------------ 

yesno  <- c("y", "n", "u", ".missing")
dyesno <- c("Yes", "No", "Unknown", "Missing")

treatment_administered  <- c(0:1, ".missing")
dtreatment_administered <- c("Yes", "No", "Missing")

facility  <- c(1:10, ".default") # define a .default key
dfacility <- c(sprintf("Facility %s", format(1:10)), "Unknown")

age_group  <- c(0, 10, 20, 30, 40, 50)
dage_group <- c("0-9", "10-19", "20-29", "30-39", "40-49", "50+")

wordlist <- data.frame(
  options = c(yesno, treatment_administered, facility, age_group),
  values  = c(dyesno, dtreatment_administered, dfacility, dage_group),
  grp = rep(c("readmission", "treatment_administered", "facility", "age_group"),
            c(4, 3, 11, 6)),
  orders  = c(1:4, 1:3, 1:11, 1:6),
  stringsAsFactors = FALSE
)

# Assigning global values ----------------------------------------

global_words <- data.frame(
  options = c("y", "n", "u", "unk", "oui", ".missing"),
  values  = c("yes", "no", "unknown", "unknown", "yes", "missing"),
  grp     = rep(".global", 6),
  orders  = rep(Inf, 6),
  stringsAsFactors = FALSE
)

wordlist <- rbind(wordlist, global_words, stringsAsFactors = FALSE)

# Generate example data ------------------------------------------
dat <- data.frame(
  # these have been defined
  readmission = sample(yesno, 50, replace = TRUE),
  treatment_administered = sample(treatment_administered, 50, replace = TRUE),
  facility = sample(c(facility[-11], LETTERS[1:3]), 50, replace = TRUE),
  age_group = sample(age_group, 50, replace = TRUE),
  # global values will catch these
  has_symptoms = sample(c(yesno, "unk", "oui"), 50, replace = TRUE),
  followup = sample(c(yesno, "unk", "oui"), 50, replace = TRUE),
  stringsAsFactors = FALSE
)
missing_data <- dat == ".missing"
dat[missing_data] <- sample(c("", NA), sum(missing_data), prob = c(0.1, 0.9), replace = TRUE)

# Clean spelling based on wordlist ------------------------------ 

wordlist # show the wordlist

#>     options      values                    grp orders
#> 1         y         Yes            readmission      1
#> 2         n          No            readmission      2
#> 3         u     Unknown            readmission      3
#> 4  .missing     Missing            readmission      4
#> 5         0         Yes treatment_administered      1
#> 6         1          No treatment_administered      2
#> 7  .missing     Missing treatment_administered      3
#> 8         1 Facility  1               facility      1
#> 9         2 Facility  2               facility      2
#> 10        3 Facility  3               facility      3
#> 11        4 Facility  4               facility      4
#> 12        5 Facility  5               facility      5
#> 13        6 Facility  6               facility      6
#> 14        7 Facility  7               facility      7
#> 15        8 Facility  8               facility      8
#> 16        9 Facility  9               facility      9
#> 17       10 Facility 10               facility     10
#> 18 .default     Unknown               facility     11
#> 19        0         0-9              age_group      1
#> 20       10       10-19              age_group      2
#> 21       20       20-29              age_group      3
#> 22       30       30-39              age_group      4
#> 23       40       40-49              age_group      5
#> 24       50         50+              age_group      6
#> 25        y         yes                .global    Inf
#> 26        n          no                .global    Inf
#> 27        u     unknown                .global    Inf
#> 28      unk     unknown                .global    Inf
#> 29      oui         yes                .global    Inf
#> 30 .missing     missing                .global    Inf

head(dat) # show the data

#>   readmission treatment_administered facility age_group has_symptoms followup
#> 1           u                      0        8        20            n        n
#> 2           u                      1        B        50            u     <NA>
#> 3           n                      0        B        40          unk      oui
#> 4        <NA>                      1        8        20          oui      unk
#> 5           y                      1        4        40            u     <NA>
#> 6        <NA>                      0        C        40          oui        n

head(clean_variable_spelling(dat, wordlists = wordlist, spelling_vars = "grp"))

#>   readmission treatment_administered    facility age_group has_symptoms
#> 1     Unknown                    Yes Facility  8     20-29           no
#> 2     Unknown                     No     Unknown       50+      unknown
#> 3          No                    Yes     Unknown     40-49      unknown
#> 4     Missing                     No Facility  8     20-29          yes
#> 5         Yes                     No Facility  4     40-49      unknown
#> 6     Missing                    Yes     Unknown     40-49          yes
#>   followup
#> 1       no
#> 2  missing
#> 3      yes
#> 4  unknown
#> 5  missing
#> 6       no

# You can ensure the order of the factors are correct by specifying 
# a column that defines order.

dat[] <- lapply(dat, as.factor)
as.list(head(dat))

#> $readmission
#> [1] u    u    n    <NA> y    <NA>
#> Levels:  n u y
#> 
#> $treatment_administered
#> [1] 0 1 0 1 1 0
#> Levels:  0 1
#> 
#> $facility
#> [1] 8 B B 8 4 C
#> Levels: 1 10 2 3 4 5 6 7 8 9 A B C
#> 
#> $age_group
#> [1] 20 50 40 20 40 40
#> Levels: 0 10 20 30 40 50
#> 
#> $has_symptoms
#> [1] n   u   unk oui u   oui
#> Levels: n oui u unk y
#> 
#> $followup
#> [1] n    <NA> oui  unk  <NA> n   
#> Levels: n oui u unk y
#> 

res <- clean_variable_spelling(dat,
                               wordlists = wordlist,
                               spelling_vars = "grp",
                               sort_by = "orders")
head(res)

#>   readmission treatment_administered    facility age_group has_symptoms
#> 1     Unknown                    Yes Facility  8     20-29           no
#> 2     Unknown                     No     Unknown       50+      unknown
#> 3          No                    Yes     Unknown     40-49      unknown
#> 4     Missing                     No Facility  8     20-29          yes
#> 5         Yes                     No Facility  4     40-49      unknown
#> 6     Missing                    Yes     Unknown     40-49          yes
#>   followup
#> 1       no
#> 2  missing
#> 3      yes
#> 4  unknown
#> 5  missing
#> 6       no

as.list(head(res))

#> $readmission
#> [1] Unknown Unknown No      Missing Yes     Missing
#> Levels: Yes No Unknown Missing
#> 
#> $treatment_administered
#> [1] Yes No  Yes No  No  Yes
#> Levels: Yes No Missing
#> 
#> $facility
#> [1] Facility  8 Unknown     Unknown     Facility  8 Facility  4 Unknown    
#> 11 Levels: Facility  1 Facility  2 Facility  3 Facility  4 ... Unknown
#> 
#> $age_group
#> [1] 20-29 50+   40-49 20-29 40-49 40-49
#> Levels: 0-9 10-19 20-29 30-39 40-49 50+
#> 
#> $has_symptoms
#> [1] no      unknown unknown yes     unknown yes    
#> Levels: yes no unknown missing
#> 
#> $followup
#> [1] no      missing yes     unknown missing no     
#> Levels: yes no unknown missing
#>