List registred epivars in a linelist object

List registred epivars in a linelist object

list_epivars(x, simple = FALSE, full_dict = FALSE)

Arguments

x
a linelist object

simple
a logical indicating if a single vector of names of present epivars should be returned (TRUE); otherwise the output will contain more information as a data.frame (FALSE, default)

full_dict
if FALSE (default), only the epivars present in the object are documented; otherwise, all the available dictionary is appended to the output

Value

if simple is TRUE, a vector of characters giving the names of available epivariables; otherwise, a data.frame giving more information about these variables

Examples

dat <- clean_data(messy_data())
ll  <- as_linelist(dat,
                  id = "id",
                  gender = "gender",
                  date_onset = "date_of_onset",
                  geo_lon = "lon",
                  geo_lat = "lat"
                 )
list_epivars(ll)

#>       epivar        column              hxl                      description
#> 1         id            id       #respondee     unique individual identifier
#> 2 date_onset date_of_onset     #date +start   date at which symptoms started
#> 3     gender        gender #indicator +type         gender of the individual
#> 4    geo_lat           lat        #geo +lat  geographic coordinate: latitude
#> 5    geo_lon           lon        #geo +lon geographic coordinate: longitude

list_epivars(ll, simple = TRUE)

#> [1] "id"         "date_onset" "gender"     "geo_lat"    "geo_lon"   

list_epivars(ll, full_dict = TRUE)

#>             epivar        column              hxl
#> 1               id            id       #respondee
#> 2       date_onset date_of_onset     #date +start
#> 3           gender        gender #indicator +type
#> 4          geo_lat           lat        #geo +lat
#> 5          geo_lon           lon        #geo +lon
#> 6  case_definition          <NA> #indicator +name
#> 7          outcome          <NA> #indicator +type
#> 8      date_report          <NA>  #date +reported
#> 9     date_outcome          <NA>       #date +end
#> 10       age_group          <NA> #indicator +type
#> 11             age          <NA>  #indicator +num
#>                                            description
#> 1                         unique individual identifier
#> 2                       date at which symptoms started
#> 3                             gender of the individual
#> 4                      geographic coordinate: latitude
#> 5                     geographic coordinate: longitude
#> 6  case type: suspected, probable, confirmed, negative
#> 7                                    recovery or death
#> 8                      date at which case was reported
#> 9              date of the outcome (recovery or death)
#> 10                age group of the individual in years
#> 11                      age of the individual in years