Reversibly rename columns in your linelist with epivars

Reversibly rename columns in your linelist with epivars

mask(x)

# S3 method for default
mask(x)

# S3 method for linelist
mask(x)

unmask(x)

# S3 method for default
unmask(x)

# S3 method for linelist
unmask(x)

# S3 method for data.frame
unmask(x)

Arguments

x
a linelist object

Value

mask() a linelist object with columns renamed by their corresponding epivars and an attribute called "masked-linelist" containing a vector of the original column names named by their corresponding epivars

unmask() a linelist object

Examples

md <- messy_data(10)
cd <- clean_data(md)
ll <- as_linelist(cd,
                  id = "id",
                  date_onset = "date_of_onset",
                  gender = "gender",
                  case_definition = "epi_case_definition",
                  geo_lon = "lon",
                  geo_lat = "lat"
                 )
ll

#> <linelist object>
#> 
#>        id date_of_onset  discharge gender epi_case_definition messy_dates
#> 1  dzoshv    2018-01-09 2018-01-19 female           suspected        <NA>
#> 2  ibcboe    2018-01-05 2018-01-15 female            probable  2018-10-18
#> 3  rzjtsb    2018-01-09 2018-01-19   male           suspected  1989-12-24
#> 4  untjny    2018-01-03 2018-01-13   male          not_a_case  2018-10-17
#> 5  aglfzi    2018-01-05 2018-01-15   male            probable  1989-12-24
#> 6  qqrxqe    2018-01-03 2018-01-13   male            probable  2018-10-18
#> 7  fdzqnh    2018-01-06 2018-01-16   male            probable        <NA>
#> 8  bszjbf    2018-01-11 2018-01-21   male          not_a_case  2001-12-01
#> 9  ixkfck    2018-01-06 2018-01-16 female           suspected  1989-12-24
#> 10 iklzik    2018-01-03 2018-01-13 female           confirmed  1989-12-24
#>         lat      lon
#> 1  13.41707 49.78313
#> 2  15.24252 49.03138
#> 3  11.69601 47.41930
#> 4  10.81973 46.40692
#> 5  11.57564 47.51096
#> 6  10.27469 47.24730
#> 7  11.99862 48.66538
#> 8  15.87695 48.73566
#> 9  11.97672 48.73080
#> 10 12.99606 47.47127

# masking epivars changes the column names
print(llm <- mask(ll))

#> <linelist object>
#> 
#>        id date_onset  discharge gender case_definition messy_dates  geo_lat
#> 1  dzoshv 2018-01-09 2018-01-19 female       suspected        <NA> 13.41707
#> 2  ibcboe 2018-01-05 2018-01-15 female        probable  2018-10-18 15.24252
#> 3  rzjtsb 2018-01-09 2018-01-19   male       suspected  1989-12-24 11.69601
#> 4  untjny 2018-01-03 2018-01-13   male      not_a_case  2018-10-17 10.81973
#> 5  aglfzi 2018-01-05 2018-01-15   male        probable  1989-12-24 11.57564
#> 6  qqrxqe 2018-01-03 2018-01-13   male        probable  2018-10-18 10.27469
#> 7  fdzqnh 2018-01-06 2018-01-16   male        probable        <NA> 11.99862
#> 8  bszjbf 2018-01-11 2018-01-21   male      not_a_case  2001-12-01 15.87695
#> 9  ixkfck 2018-01-06 2018-01-16 female       suspected  1989-12-24 11.97672
#> 10 iklzik 2018-01-03 2018-01-13 female       confirmed  1989-12-24 12.99606
#>     geo_lon
#> 1  49.78313
#> 2  49.03138
#> 3  47.41930
#> 4  46.40692
#> 5  47.51096
#> 6  47.24730
#> 7  48.66538
#> 8  48.73566
#> 9  48.73080
#> 10 47.47127

# unmasking reverses the operation
unmask(llm)

#> <linelist object>
#> 
#>        id date_of_onset  discharge gender epi_case_definition messy_dates
#> 1  dzoshv    2018-01-09 2018-01-19 female           suspected        <NA>
#> 2  ibcboe    2018-01-05 2018-01-15 female            probable  2018-10-18
#> 3  rzjtsb    2018-01-09 2018-01-19   male           suspected  1989-12-24
#> 4  untjny    2018-01-03 2018-01-13   male          not_a_case  2018-10-17
#> 5  aglfzi    2018-01-05 2018-01-15   male            probable  1989-12-24
#> 6  qqrxqe    2018-01-03 2018-01-13   male            probable  2018-10-18
#> 7  fdzqnh    2018-01-06 2018-01-16   male            probable        <NA>
#> 8  bszjbf    2018-01-11 2018-01-21   male          not_a_case  2001-12-01
#> 9  ixkfck    2018-01-06 2018-01-16 female           suspected  1989-12-24
#> 10 iklzik    2018-01-03 2018-01-13 female           confirmed  1989-12-24
#>         lat      lon
#> 1  13.41707 49.78313
#> 2  15.24252 49.03138
#> 3  11.69601 47.41930
#> 4  10.81973 46.40692
#> 5  11.57564 47.51096
#> 6  10.27469 47.24730
#> 7  11.99862 48.66538
#> 8  15.87695 48.73566
#> 9  11.97672 48.73080
#> 10 12.99606 47.47127

# you can safely drop columns
unmask(llm[grepl("gender|geo", names(llm))])

#> <linelist object>
#> 
#>    gender      lat      lon
#> 1  female 13.41707 49.78313
#> 2  female 15.24252 49.03138
#> 3    male 11.69601 47.41930
#> 4    male 10.81973 46.40692
#> 5    male 11.57564 47.51096
#> 6    male 10.27469 47.24730
#> 7    male 11.99862 48.66538
#> 8    male 15.87695 48.73566
#> 9  female 11.97672 48.73080
#> 10 female 12.99606 47.47127