Add epivars to a linelist object

These functions can be used to change the epivars registred in a linelist object.

set_epivars(x, ...)

reset_epivars(x, ...)

# S3 method for linelist
set_epivars(x, ...)

# S3 method for linelist
reset_epivars(x, ...)

# S3 method for data.frame
set_epivars(x, ...)

# S3 method for data.frame
reset_epivars(x, ...)

Arguments

x
a linelist object

...
a series of column names named after epivars

Value

a linelist object with new epivars ammended

Details

set_epivars adds new epivars to the existing ones.

reset_epivars replaces the existing epivars with new ones.

Examples

# define some toy data
cd <- clean_data(messy_data())

# Use as_linelist to define your epivars ---------------------
ll <- as_linelist(cd,
  "gender"          = "gender",
  "id"              = "id",
  "geo_lon"         = "lon",
  "geo_lat"         = "lat",
  "case_definition" = "epi_case_definition"
)
list_epivars(ll)

#>            epivar              column              hxl
#> 1              id                  id       #respondee
#> 2          gender              gender #indicator +type
#> 3 case_definition epi_case_definition #indicator +name
#> 4         geo_lat                 lat        #geo +lat
#> 5         geo_lon                 lon        #geo +lon
#>                                           description
#> 1                        unique individual identifier
#> 2                            gender of the individual
#> 3 case type: suspected, probable, confirmed, negative
#> 4                     geographic coordinate: latitude
#> 5                    geographic coordinate: longitude

# add a new epivar -------------------------------------------
ll <- set_epivars(ll, "date_onset" = "date_of_onset")
list_epivars(ll)

#>            epivar              column              hxl
#> 1              id                  id       #respondee
#> 2      date_onset       date_of_onset     #date +start
#> 3          gender              gender #indicator +type
#> 4 case_definition epi_case_definition #indicator +name
#> 5         geo_lat                 lat        #geo +lat
#> 6         geo_lon                 lon        #geo +lon
#>                                           description
#> 1                        unique individual identifier
#> 2                      date at which symptoms started
#> 3                            gender of the individual
#> 4 case type: suspected, probable, confirmed, negative
#> 5                     geographic coordinate: latitude
#> 6                    geographic coordinate: longitude

# change an epivar column ------------------------------------
# 
# in this case, we want to update the date of onset to reflect the number
# of days from the start 
ll$days_from_start <- as.integer(get_epivars(ll, "date_onset") - as.Date("2018-01-23"))
ll <- set_epivars(ll, "date_onset" = "days_from_start")
list_epivars(ll)

#>            epivar              column              hxl
#> 1              id                  id       #respondee
#> 2          gender              gender #indicator +type
#> 3 case_definition epi_case_definition #indicator +name
#> 4         geo_lat                 lat        #geo +lat
#> 5         geo_lon                 lon        #geo +lon
#> 6      date_onset     days_from_start     #date +start
#>                                           description
#> 1                        unique individual identifier
#> 2                            gender of the individual
#> 3 case type: suspected, probable, confirmed, negative
#> 4                     geographic coordinate: latitude
#> 5                    geographic coordinate: longitude
#> 6                      date at which symptoms started

# Remove an epivar label -------------------------------------
ll <- set_epivars(ll, "date_onset" = NULL)
list_epivars(ll)

#>            epivar              column              hxl
#> 1              id                  id       #respondee
#> 2          gender              gender #indicator +type
#> 3 case_definition epi_case_definition #indicator +name
#> 4         geo_lat                 lat        #geo +lat
#> 5         geo_lon                 lon        #geo +lon
#>                                           description
#> 1                        unique individual identifier
#> 2                            gender of the individual
#> 3 case type: suspected, probable, confirmed, negative
#> 4                     geographic coordinate: latitude
#> 5                    geographic coordinate: longitude