Changes tags of a linelist object

This function changes the tags of a linelist object, using the same syntax as the constructor make_linelist(). If some of the default tags are missing, they will be added to the final object.

set_tags(x, ..., allow_extra = FALSE)

Arguments

x

a data.frame or a tibble containing case line list data, with cases in rows and variables in columns

...

a series of tags provided as tag_name = "column_name", where tag_name indicates any of the known variables listed in 'Details'; alternatively, a named list of variables to be tagged, where names indicate the types of variable (to be selected from tags_names()), and values indicate their name in the input data.frame; see details for a list of known variable types and their expected content

allow_extra

a logical indicating if additional data tags not currently recognized by linelist should be allowed; if FALSE, unknown tags will trigger an error

Value

The function returns a linelist object.

See also

Author

Thibaut Jombart thibaut@data.org

Examples

if (require(outbreaks)) {
  ## create a linelist
  x <- make_linelist(measles_hagelloch_1861, date_onset = "date_of_rash")
  tags(x)

  ## add new tags and fix an existing one
  x <- set_tags(x, age = "age",
                gender = "gender",
                date_onset = "date_of_prodrome")
  tags(x)

  ## add non-default tags using allow_extra
  x <- set_tags(x, severe = "complications", allow_extra = TRUE)
  tags(x)

  ## remove tags by setting them to NULL
  old_tags <- tags(x)
  x <- set_tags(x, age = NULL, gender = NULL)
  tags(x)

  ## setting tags providing a list (used to restore old tags here)
  x <- set_tags(x, old_tags)
  tags(x)
  
}
#> $date_onset
#> [1] "date_of_prodrome"
#> 
#> $gender
#> [1] "gender"
#> 
#> $age
#> [1] "age"
#> 
#> $severe
#> [1] "complications"
#>