Subsetting of linelist objets

The [] operator for linelist objects behaves like for regular data.frame or tibble, but checks that tagged variables are not lost, and takes the appropriate action (warning, error, or ignore, depending on the value of lost_action) if this is the case.

# S3 method for linelist
[(x, i, j, drop = FALSE, lost_action = "warning")

Arguments

x

a linelist object

i

a vector of integer or logical to subset the rows of the linelist

j

a vector of character, integer, or logical to subset the columns of the linelist

drop

a logical indicating if, when a single column is selected, the data.frame class should be dropped to return a simple vector, in which case the linelist class is lost as well; defaults to FALSE

lost_action

a character indicating the behaviour to adopt when some of the tagged variables are dropped through the subsetting process; can be "warning" (default), "error", or "none".

Value

If no drop is happening, a linelist. Otherwise an atomic vector.

Author

Thibaut Jombart thibaut@data.org