Checks the content of a linelist object

This function evalutes the validity of a linelist object by checking the object class, its tags, and the types of the tagged variables. It combines validations checks made by validate_types and validate_tags. See 'Details' section for more information on the checks performed.

validate_linelist(x, allow_extra = FALSE, ref_types = tags_types())

Arguments

x

a linelist object

allow_extra

a logical indicating if additional data tags not currently recognized by linelist should be allowed; if FALSE, unknown tags will trigger an error

ref_types

a list providing allowed types for all tags, as returned by tags_types

Value

If checks pass, a linelist object; otherwise issues an error.

Details

The following checks are performed:

x is a linelist object

x has a well-formed tags attribute

all default tags are present (even if NULL)

all tagged variables correspond to existing columns

all tagged variables have an acceptable class

(optional) x has no extra tag beyond the default tags

See also

Author

Thibaut Jombart thibaut@data.org

Examples

if (require(outbreaks) && require(dplyr) && require(magrittr)) {

  ## create a valid linelist
  x <- measles_hagelloch_1861 %>%
    tibble() %>% 
    make_linelist(id = "case_ID",
                  date_onset = "date_of_prodrome",
                  age = "age",
                  gender = "gender")
  x

  ## validation
  validate_linelist(x)

  ## create an invalid linelist - onset date is a factor
  x <- measles_hagelloch_1861 %>%
    tibble() %>% 
    make_linelist(id = "case_ID",
                  date_onset = "gender",
                  age = "age")
  x

  ## the below issues an error
  # validate_linelist(x)
}
#> 
#> // linelist object
#> # A tibble: 188 × 12
#>    case_ID infector date_of_prodrome date_of_rash date_of_death   age gender
#>      <int>    <int> <date>           <date>       <date>        <dbl> <fct> 
#>  1       1       45 1861-11-21       1861-11-25   NA                7 f     
#>  2       2       45 1861-11-23       1861-11-27   NA                6 f     
#>  3       3      172 1861-11-28       1861-12-02   NA                4 f     
#>  4       4      180 1861-11-27       1861-11-28   NA               13 m     
#>  5       5       45 1861-11-22       1861-11-27   NA                8 f     
#>  6       6      180 1861-11-26       1861-11-29   NA               12 m     
#>  7       7       42 1861-11-24       1861-11-28   NA                6 m     
#>  8       8       45 1861-11-21       1861-11-26   NA               10 m     
#>  9       9      182 1861-11-26       1861-11-30   NA               13 m     
#> 10      10       45 1861-11-21       1861-11-25   NA                7 f     
#> # … with 178 more rows, and 5 more variables: family_ID <int>, class <fct>,
#> #   complications <fct>, x_loc <dbl>, y_loc <dbl>
#> 
#> // tags: id:case_ID, date_onset:gender, age:age