Accessor functfons

This set of fucntions defines helper functions for accessing pre-defined epivars in your linelist object.

get_epivars(x, ..., simplify = TRUE)

id(x)

date_onset(x)

date_report(x)

date_outcome(x)

case_definition(x)

outcome(x)

gender(x)

age(x)

age_group(x)

geo(x)

Arguments

x
a linelist object

...
epivars to be used; can be provided as characters, integers, or logical; integers and logicals are to be matched against the vector of epi variables returned by list_epivars(x, simple = TRUE). If nothing is provided, all epivars are returned.

simplify
if TRUE (default) and a single epivar is specified, it will be converted to a vector. Otherwise, a data frame is returned.

Value

a vector if the epivar can be represented by a single column, a data frame otherwise (e.g. a location column).

Note

For the geo() accessor, which necessarily contains two columns for longitude and latitude data, when simplify = TRUE, a matrix will be returned.

Examples

## make toy data
toy_data <- messy_data()
cleaned_data <- clean_data(toy_data)
ll <- as_linelist(cleaned_data,
                  id = "id",
                  date_onset = "date_of_onset",
                  gender = "gender",
                  geo_lon = "lon",
                  geo_lat = "lat"
                 )

## general purpose accessor
list_epivars(ll, simple = TRUE)

#> [1] "id"         "date_onset" "gender"     "geo_lat"    "geo_lon"   

get_epivars(ll) # all epi variable

#>        id date_of_onset gender       lat      lon
#> 1  vhoipe    2018-01-02 female 12.225573 50.03937
#> 2  yoohlj    2018-01-10 female 11.429135 48.44945
#> 3  amkajo    2018-01-08 female 10.886526 49.39181
#> 4  wknpgh    2018-01-03 female 11.408917 48.42657
#> 5  mxkfrb    2018-01-02   male  9.487449 48.10758
#> 6  zdnvsg    2018-01-06 female 11.618924 48.02229
#> 7  ovamuv    2018-01-06   male 11.882916 48.60361
#> 8  kfkrnr    2018-01-04   male 11.926673 47.73735
#> 9  nvswax    2018-01-09 female 13.454254 47.47174
#> 10 znjupq    2018-01-09 female 14.956910 48.19215
#> 11 wogxwg    2018-01-10   male 12.582235 46.85380
#> 12 kudfer    2018-01-03 female 10.201179 48.84618
#> 13 wyoprg    2018-01-02 female 13.517075 48.08172
#> 14 bxbzpe    2018-01-05   male 12.116401 46.69488
#> 15 xapgxv    2018-01-06 female 14.137200 47.05509
#> 16 sfsfli    2018-01-03 female 17.253701 48.45434
#> 17 ogmxzv    2018-01-06 female 13.849717 47.14480
#> 18 xpuzsf    2018-01-02   male  9.631437 47.71310
#> 19 hswgay    2018-01-05 female 13.498804 48.89496
#> 20 ltrkjt    2018-01-11 female 15.145677 48.06730

get_epivars(ll, "id", "date_onset", "gender") # named epi variables

#>        id date_of_onset gender
#> 1  vhoipe    2018-01-02 female
#> 2  yoohlj    2018-01-10 female
#> 3  amkajo    2018-01-08 female
#> 4  wknpgh    2018-01-03 female
#> 5  mxkfrb    2018-01-02   male
#> 6  zdnvsg    2018-01-06 female
#> 7  ovamuv    2018-01-06   male
#> 8  kfkrnr    2018-01-04   male
#> 9  nvswax    2018-01-09 female
#> 10 znjupq    2018-01-09 female
#> 11 wogxwg    2018-01-10   male
#> 12 kudfer    2018-01-03 female
#> 13 wyoprg    2018-01-02 female
#> 14 bxbzpe    2018-01-05   male
#> 15 xapgxv    2018-01-06 female
#> 16 sfsfli    2018-01-03 female
#> 17 ogmxzv    2018-01-06 female
#> 18 xpuzsf    2018-01-02   male
#> 19 hswgay    2018-01-05 female
#> 20 ltrkjt    2018-01-11 female

date_onset(ll)

#>  [1] "2018-01-02" "2018-01-10" "2018-01-08" "2018-01-03" "2018-01-02"
#>  [6] "2018-01-06" "2018-01-06" "2018-01-04" "2018-01-09" "2018-01-09"
#> [11] "2018-01-10" "2018-01-03" "2018-01-02" "2018-01-05" "2018-01-06"
#> [16] "2018-01-03" "2018-01-06" "2018-01-02" "2018-01-05" "2018-01-11"

id(ll)

#>  [1] vhoipe yoohlj amkajo wknpgh mxkfrb zdnvsg ovamuv kfkrnr nvswax znjupq
#> [11] wogxwg kudfer wyoprg bxbzpe xapgxv sfsfli ogmxzv xpuzsf hswgay ltrkjt
#> 20 Levels: amkajo bxbzpe hswgay kfkrnr kudfer ltrkjt mxkfrb nvswax ... znjupq

gender(ll)

#>  [1] female female female female male   female male   male   female female
#> [11] male   female female male   female female female male   female female
#> Levels: female male

geo(ll) # a matrix with two columns

#>         lon       lat
#> 1  50.03937 12.225573
#> 2  48.44945 11.429135
#> 3  49.39181 10.886526
#> 4  48.42657 11.408917
#> 5  48.10758  9.487449
#> 6  48.02229 11.618924
#> 7  48.60361 11.882916
#> 8  47.73735 11.926673
#> 9  47.47174 13.454254
#> 10 48.19215 14.956910
#> 11 46.85380 12.582235
#> 12 48.84618 10.201179
#> 13 48.08172 13.517075
#> 14 46.69488 12.116401
#> 15 47.05509 14.137200
#> 16 48.45434 17.253701
#> 17 47.14480 13.849717
#> 18 47.71310  9.631437
#> 19 48.89496 13.498804
#> 20 48.06730 15.145677

get_epivars(ll, "geo_lon", "geo_lat", simplify = FALSE) # a data frame with two columns

#>         lon       lat
#> 1  50.03937 12.225573
#> 2  48.44945 11.429135
#> 3  49.39181 10.886526
#> 4  48.42657 11.408917
#> 5  48.10758  9.487449
#> 6  48.02229 11.618924
#> 7  48.60361 11.882916
#> 8  47.73735 11.926673
#> 9  47.47174 13.454254
#> 10 48.19215 14.956910
#> 11 46.85380 12.582235
#> 12 48.84618 10.201179
#> 13 48.08172 13.517075
#> 14 46.69488 12.116401
#> 15 47.05509 14.137200
#> 16 48.45434 17.253701
#> 17 47.14480 13.849717
#> 18 47.71310  9.631437
#> 19 48.89496 13.498804
#> 20 48.06730 15.145677

# epivars that haven't been defined for the data set will return an error
try(date_report(ll))

#> Error in get_epivars(x, "date_report") : 
#>   The following epivars were not found in the data:
#> date_report