Clean a data.frame

This function applies several cleaning procedures to an input data.frame, by standardising variable names, labels used categorical variables (characters of factors), and setting dates to Date objects. Optionally, an intelligent date search can be used on character strings to extract dates from various formats mixed with other text. See details for more information.

clean_data(x, sep = "_", force_Date = TRUE, guess_dates = TRUE,
  error_tolerance = 0.5, wordlists = NULL, spelling_vars = 3,
  sort_by = NULL, warn_spelling = FALSE, protect = FALSE, ...)

Arguments

x
a data.frame

sep
The separator used between words, and defaults to the underscore _.

force_Date
a logical or integer vector indicating the columns . If logical, indicating if POSIXct and POSIXlt objects should be converted to Date objects; defaults to TRUE; you should use this if your dates are only precise to the day (i.e. no time information within days).

guess_dates
a logical or integer vector indicating which columns should be guessed , assuming these columns store character strings or factors; this feature is experimental; see guess_dates() for more information.

error_tolerance
a number between 0 and 1 indicating the proportion of entries which cannot be identified as dates to be tolerated; if this proportion is exceeded, the original vector is returned, and a message is issued; defaults to 0.1 (10 percent)

wordlists
a data frame or named list of data frames with at least two columns defining the word list to be used. If this is a data frame, a third column must be present to split the wordlists by column in x (see spelling_vars).

spelling_vars
character or integer. If wordlists is a data frame, then this column in defines the columns in x corresponding to each section of the wordlists data frame. This defaults to 3, indicating the third column is to be used.

sort_by
a character the column to be used for sorting the values in each data frame. If the incoming variables are factors, this determines how the resulting factors will be sorted.

warn_spelling
if TRUE, errors and warnings from clean_spelling() will be aggregated and presented for each column that issues them. The default value is FALSE, which means that all errors and warnings will be ignored.

protect
a logical or numeric vector defining the columns to protect from any manipulation. Note: columns in protect will override any columns in either force_Date or guess_dates.

...
further arguments passed on to guess_dates()

Value

A data.frame with standardised labels for characters and factors.

Note

Creating your wordlist
When creating the wordlist for clean_variable_spelling(), it's important to remember that the data will first be cleaned with clean_variable_labels(), which will remove any capitalisation, accents, and replace all punctuation and spaces with "_".
See also

Examples

## make toy data
toy_data <- messy_data(20)

## show data
toy_data

#>       'ID Date of Onset. DisCharge.. GENDER_  Épi.Case_définition
#> 1  qvjqle     2018-01-04  14/01/2018   FEMALE           confirmed
#> 2  ndqenc     2018-01-09  19/01/2018   female          not a case
#> 3  nbvhgk     2018-01-06  16/01/2018     MALE         suspected  
#> 4  qyiqvd     2018-01-08  18/01/2018     male           suspected
#> 5  zcgqng     2018-01-09  19/01/2018   Female           confirmed
#> 6  zrcqwy     2018-01-02  12/01/2018   FEMALE           confirmed
#> 7  kmjgyj     2018-01-11  21/01/2018     MALE           suspected
#> 8  fsefvi     2018-01-04  14/01/2018     male           Confirmed
#> 9  badaty     2018-01-11  21/01/2018   FEMALE            PROBABLE
#> 10 zkhmba     2018-01-07  17/01/2018     Male            PROBABLE
#> 11 mggogq     2018-01-11  21/01/2018     MALE           suspected
#> 12 iyxcto     2018-01-06  16/01/2018     male           confirmed
#> 13 ssblvo     2018-01-05  15/01/2018     male           confirmed
#> 14 pwsylh     2018-01-06  16/01/2018   Female         suspected  
#> 15 uqekak     2018-01-11  21/01/2018   female            probable
#> 16 tugtal     2018-01-04  14/01/2018     male          not a case
#> 17 nlvxrc     2018-01-04  14/01/2018     MALE          Not.a.Case
#> 18 fmsnwh     2018-01-10  20/01/2018     male            PROBABLE
#> 19 frimdr     2018-01-03  13/01/2018   FEMALE            PROBABLE
#> 20 ivoybq     2018-01-10  20/01/2018     MALE           suspected
#>           messy/dates       lat      lon
#> 1          2018 10 19 16.558883 47.37041
#> 2     // 24//12//1989 13.773262 47.19227
#> 3              female 11.162610 47.13951
#> 4          2018 10 19  9.831327 45.83076
#> 5                male 12.831882 46.62416
#> 6          2018_10_17  8.829858 47.50687
#> 7                male 13.007136 47.41835
#> 8                male 12.288458 47.83277
#> 9          2018-10-18 15.292720 48.48599
#> 10 that's 24/12/1989! 12.557623 46.66660
#> 11               <NA> 15.036358 47.73803
#> 12             female 12.472561 48.65239
#> 13         2018 10 19 16.317085 48.74885
#> 14               male 11.451826 48.89656
#> 15             female 11.152124 49.48930
#> 16         2018 10 19 12.448933 47.34060
#> 17         2018_10_17 11.813201 48.53728
#> 18    // 24//12//1989 12.755428 48.74680
#> 19 that's 24/12/1989! 15.359568 49.89632
#> 20         01-12-2001 14.282075 45.93993

## clean variable names, store in new object, show results
clean_data <- clean_data(toy_data, error_tolerance = 0.1)
clean_data

#>        id date_of_onset  discharge gender epi_case_definition       messy_dates
#> 1  qvjqle    2018-01-04 2018-01-14 female           confirmed        2018_10_19
#> 2  ndqenc    2018-01-09 2018-01-19 female          not_a_case        24_12_1989
#> 3  nbvhgk    2018-01-06 2018-01-16   male           suspected            female
#> 4  qyiqvd    2018-01-08 2018-01-18   male           suspected        2018_10_19
#> 5  zcgqng    2018-01-09 2018-01-19 female           confirmed              male
#> 6  zrcqwy    2018-01-02 2018-01-12 female           confirmed        2018_10_17
#> 7  kmjgyj    2018-01-11 2018-01-21   male           suspected              male
#> 8  fsefvi    2018-01-04 2018-01-14   male           confirmed              male
#> 9  badaty    2018-01-11 2018-01-21 female            probable        2018_10_18
#> 10 zkhmba    2018-01-07 2018-01-17   male            probable that_s_24_12_1989
#> 11 mggogq    2018-01-11 2018-01-21   male           suspected              <NA>
#> 12 iyxcto    2018-01-06 2018-01-16   male           confirmed            female
#> 13 ssblvo    2018-01-05 2018-01-15   male           confirmed        2018_10_19
#> 14 pwsylh    2018-01-06 2018-01-16 female           suspected              male
#> 15 uqekak    2018-01-11 2018-01-21 female            probable            female
#> 16 tugtal    2018-01-04 2018-01-14   male          not_a_case        2018_10_19
#> 17 nlvxrc    2018-01-04 2018-01-14   male          not_a_case        2018_10_17
#> 18 fmsnwh    2018-01-10 2018-01-20   male            probable        24_12_1989
#> 19 frimdr    2018-01-03 2018-01-13 female            probable that_s_24_12_1989
#> 20 ivoybq    2018-01-10 2018-01-20   male           suspected        01_12_2001
#>          lat      lon
#> 1  16.558883 47.37041
#> 2  13.773262 47.19227
#> 3  11.162610 47.13951
#> 4   9.831327 45.83076
#> 5  12.831882 46.62416
#> 6   8.829858 47.50687
#> 7  13.007136 47.41835
#> 8  12.288458 47.83277
#> 9  15.292720 48.48599
#> 10 12.557623 46.66660
#> 11 15.036358 47.73803
#> 12 12.472561 48.65239
#> 13 16.317085 48.74885
#> 14 11.451826 48.89656
#> 15 11.152124 49.48930
#> 16 12.448933 47.34060
#> 17 11.813201 48.53728
#> 18 12.755428 48.74680
#> 19 15.359568 49.89632
#> 20 14.282075 45.93993

clean_data2 <- clean_data(toy_data, error_tolerance = 0.8)
clean_data2

#>        id date_of_onset  discharge gender epi_case_definition messy_dates
#> 1  qvjqle    2018-01-04 2018-01-14 female           confirmed  2018-10-19
#> 2  ndqenc    2018-01-09 2018-01-19 female          not_a_case  1989-12-24
#> 3  nbvhgk    2018-01-06 2018-01-16   male           suspected        <NA>
#> 4  qyiqvd    2018-01-08 2018-01-18   male           suspected  2018-10-19
#> 5  zcgqng    2018-01-09 2018-01-19 female           confirmed        <NA>
#> 6  zrcqwy    2018-01-02 2018-01-12 female           confirmed  2018-10-17
#> 7  kmjgyj    2018-01-11 2018-01-21   male           suspected        <NA>
#> 8  fsefvi    2018-01-04 2018-01-14   male           confirmed        <NA>
#> 9  badaty    2018-01-11 2018-01-21 female            probable  2018-10-18
#> 10 zkhmba    2018-01-07 2018-01-17   male            probable  1989-12-24
#> 11 mggogq    2018-01-11 2018-01-21   male           suspected        <NA>
#> 12 iyxcto    2018-01-06 2018-01-16   male           confirmed        <NA>
#> 13 ssblvo    2018-01-05 2018-01-15   male           confirmed  2018-10-19
#> 14 pwsylh    2018-01-06 2018-01-16 female           suspected        <NA>
#> 15 uqekak    2018-01-11 2018-01-21 female            probable        <NA>
#> 16 tugtal    2018-01-04 2018-01-14   male          not_a_case  2018-10-19
#> 17 nlvxrc    2018-01-04 2018-01-14   male          not_a_case  2018-10-17
#> 18 fmsnwh    2018-01-10 2018-01-20   male            probable  1989-12-24
#> 19 frimdr    2018-01-03 2018-01-13 female            probable  1989-12-24
#> 20 ivoybq    2018-01-10 2018-01-20   male           suspected  2001-12-01
#>          lat      lon
#> 1  16.558883 47.37041
#> 2  13.773262 47.19227
#> 3  11.162610 47.13951
#> 4   9.831327 45.83076
#> 5  12.831882 46.62416
#> 6   8.829858 47.50687
#> 7  13.007136 47.41835
#> 8  12.288458 47.83277
#> 9  15.292720 48.48599
#> 10 12.557623 46.66660
#> 11 15.036358 47.73803
#> 12 12.472561 48.65239
#> 13 16.317085 48.74885
#> 14 11.451826 48.89656
#> 15 11.152124 49.48930
#> 16 12.448933 47.34060
#> 17 11.813201 48.53728
#> 18 12.755428 48.74680
#> 19 15.359568 49.89632
#> 20 14.282075 45.93993

## clean variable names, but keep our "messy/dates" column
to_protect <- names(toy_data) %in% "messy/dates"
clean_data3 <- clean_data(toy_data,
                          error_tolerance = 0.8,
                          protect = to_protect
                         )
clean_data3

#>        id date_of_onset  discharge gender epi_case_definition
#> 1  qvjqle    2018-01-04 2018-01-14 female           confirmed
#> 2  ndqenc    2018-01-09 2018-01-19 female          not_a_case
#> 3  nbvhgk    2018-01-06 2018-01-16   male           suspected
#> 4  qyiqvd    2018-01-08 2018-01-18   male           suspected
#> 5  zcgqng    2018-01-09 2018-01-19 female           confirmed
#> 6  zrcqwy    2018-01-02 2018-01-12 female           confirmed
#> 7  kmjgyj    2018-01-11 2018-01-21   male           suspected
#> 8  fsefvi    2018-01-04 2018-01-14   male           confirmed
#> 9  badaty    2018-01-11 2018-01-21 female            probable
#> 10 zkhmba    2018-01-07 2018-01-17   male            probable
#> 11 mggogq    2018-01-11 2018-01-21   male           suspected
#> 12 iyxcto    2018-01-06 2018-01-16   male           confirmed
#> 13 ssblvo    2018-01-05 2018-01-15   male           confirmed
#> 14 pwsylh    2018-01-06 2018-01-16 female           suspected
#> 15 uqekak    2018-01-11 2018-01-21 female            probable
#> 16 tugtal    2018-01-04 2018-01-14   male          not_a_case
#> 17 nlvxrc    2018-01-04 2018-01-14   male          not_a_case
#> 18 fmsnwh    2018-01-10 2018-01-20   male            probable
#> 19 frimdr    2018-01-03 2018-01-13 female            probable
#> 20 ivoybq    2018-01-10 2018-01-20   male           suspected
#>           messy/dates       lat      lon
#> 1          2018 10 19 16.558883 47.37041
#> 2     // 24//12//1989 13.773262 47.19227
#> 3              female 11.162610 47.13951
#> 4          2018 10 19  9.831327 45.83076
#> 5                male 12.831882 46.62416
#> 6          2018_10_17  8.829858 47.50687
#> 7                male 13.007136 47.41835
#> 8                male 12.288458 47.83277
#> 9          2018-10-18 15.292720 48.48599
#> 10 that's 24/12/1989! 12.557623 46.66660
#> 11               <NA> 15.036358 47.73803
#> 12             female 12.472561 48.65239
#> 13         2018 10 19 16.317085 48.74885
#> 14               male 11.451826 48.89656
#> 15             female 11.152124 49.48930
#> 16         2018 10 19 12.448933 47.34060
#> 17         2018_10_17 11.813201 48.53728
#> 18    // 24//12//1989 12.755428 48.74680
#> 19 that's 24/12/1989! 15.359568 49.89632
#> 20         01-12-2001 14.282075 45.93993

## Using a wordlist  -------------------------------

# location data with mis-spellings, French, and English.
messy_locations <- c("hopsital", "h\u00f4pital", "hospital",
                     "m\u00e9dical", "clinic",
                     "feild", "field")
toy_data$location <- factor(sample(messy_locations, 20, replace = TRUE))

# show data 
toy_data$location

#>  [1] médical  clinic   hospital hôpital  hospital clinic   hopsital médical 
#>  [9] hospital clinic   hospital feild    hopsital hospital clinic   field   
#> [17] hospital hospital field    field   
#> Levels: clinic feild field hôpital hopsital hospital médical

# add a wordlist
wordlist <- data.frame(
  from  = c("hopsital", "hopital",  "medical", "feild"),
  to    = c("hospital", "hospital", "clinic",  "field"),
  variables = rep("location", 4),
  stringsAsFactors = FALSE
)

clean_data4 <- clean_data(toy_data,
                          wordlists     = wordlist,
                          spelling_vars = "variables"
                         )
clean_data4

#>        id date_of_onset  discharge gender epi_case_definition messy_dates
#> 1  qvjqle    2018-01-04 2018-01-14 female           confirmed  2018-10-19
#> 2  ndqenc    2018-01-09 2018-01-19 female          not_a_case  1989-12-24
#> 3  nbvhgk    2018-01-06 2018-01-16   male           suspected        <NA>
#> 4  qyiqvd    2018-01-08 2018-01-18   male           suspected  2018-10-19
#> 5  zcgqng    2018-01-09 2018-01-19 female           confirmed        <NA>
#> 6  zrcqwy    2018-01-02 2018-01-12 female           confirmed  2018-10-17
#> 7  kmjgyj    2018-01-11 2018-01-21   male           suspected        <NA>
#> 8  fsefvi    2018-01-04 2018-01-14   male           confirmed        <NA>
#> 9  badaty    2018-01-11 2018-01-21 female            probable  2018-10-18
#> 10 zkhmba    2018-01-07 2018-01-17   male            probable  1989-12-24
#> 11 mggogq    2018-01-11 2018-01-21   male           suspected        <NA>
#> 12 iyxcto    2018-01-06 2018-01-16   male           confirmed        <NA>
#> 13 ssblvo    2018-01-05 2018-01-15   male           confirmed  2018-10-19
#> 14 pwsylh    2018-01-06 2018-01-16 female           suspected        <NA>
#> 15 uqekak    2018-01-11 2018-01-21 female            probable        <NA>
#> 16 tugtal    2018-01-04 2018-01-14   male          not_a_case  2018-10-19
#> 17 nlvxrc    2018-01-04 2018-01-14   male          not_a_case  2018-10-17
#> 18 fmsnwh    2018-01-10 2018-01-20   male            probable  1989-12-24
#> 19 frimdr    2018-01-03 2018-01-13 female            probable  1989-12-24
#> 20 ivoybq    2018-01-10 2018-01-20   male           suspected  2001-12-01
#>          lat      lon location
#> 1  16.558883 47.37041   clinic
#> 2  13.773262 47.19227   clinic
#> 3  11.162610 47.13951 hospital
#> 4   9.831327 45.83076 hospital
#> 5  12.831882 46.62416 hospital
#> 6   8.829858 47.50687   clinic
#> 7  13.007136 47.41835 hospital
#> 8  12.288458 47.83277   clinic
#> 9  15.292720 48.48599 hospital
#> 10 12.557623 46.66660   clinic
#> 11 15.036358 47.73803 hospital
#> 12 12.472561 48.65239    field
#> 13 16.317085 48.74885 hospital
#> 14 11.451826 48.89656 hospital
#> 15 11.152124 49.48930   clinic
#> 16 12.448933 47.34060    field
#> 17 11.813201 48.53728 hospital
#> 18 12.755428 48.74680 hospital
#> 19 15.359568 49.89632    field
#> 20 14.282075 45.93993    field

clean_data4$location

#>  [1] clinic   clinic   hospital hospital hospital clinic   hospital clinic  
#>  [9] hospital clinic   hospital field    hospital hospital clinic   field   
#> [17] hospital hospital field    field   
#> Levels: hospital clinic field