Clean a data.frame

This function applies several cleaning procedures to an input data.frame, by standardising variable names, labels used categorical variables (characters of factors), and setting dates to Date objects. Optionally, an intelligent date search can be used on character strings to extract dates from various formats mixed with other text. See details for more information.

clean_data(x, sep = "_", force_Date = TRUE, guess_dates = TRUE,
  error_tolerance = 0.5, wordlists = NULL, spelling_vars = 3,
  sort_by = NULL, warn_spelling = FALSE, protect = FALSE, ...)

Arguments

x
a data.frame

sep
The separator used between words, and defaults to the underscore _.

force_Date
a logical or integer vector indicating the columns . If logical, indicating if POSIXct and POSIXlt objects should be converted to Date objects; defaults to TRUE; you should use this if your dates are only precise to the day (i.e. no time information within days).

guess_dates
a logical or integer vector indicating which columns should be guessed , assuming these columns store character strings or factors; this feature is experimental; see guess_dates() for more information.

error_tolerance
a number between 0 and 1 indicating the proportion of entries which cannot be identified as dates to be tolerated; if this proportion is exceeded, the original vector is returned, and a message is issued; defaults to 0.1 (10 percent)

wordlists
a data frame or named list of data frames with at least two columns defining the word list to be used. If this is a data frame, a third column must be present to split the wordlists by column in x (see spelling_vars).

spelling_vars
character or integer. If wordlists is a data frame, then this column in defines the columns in x corresponding to each section of the wordlists data frame. This defaults to 3, indicating the third column is to be used.

sort_by
a character the column to be used for sorting the values in each data frame. If the incoming variables are factors, this determines how the resulting factors will be sorted.

warn_spelling
if TRUE, errors and warnings from clean_spelling() will be aggregated and presented for each column that issues them. The default value is FALSE, which means that all errors and warnings will be ignored.

protect
a logical or numeric vector defining the columns to protect from any manipulation. Note: columns in protect will override any columns in either force_Date or guess_dates.

...
further arguments passed on to guess_dates()

Value

A data.frame with standardised labels for characters and factors.

Note

Creating your wordlist
When creating the wordlist for clean_variable_spelling(), it's important to remember that the data will first be cleaned with clean_variable_labels(), which will remove any capitalisation, accents, and replace all punctuation and spaces with "_".
See also

Examples

## make toy data
toy_data <- messy_data(20)

## show data
toy_data

#>       'ID Date of Onset. DisCharge.. GENDER_  Épi.Case_définition
#> 1  fnkgjs     2018-01-11  21/01/2018   FEMALE           confirmed
#> 2  vbjzfx     2018-01-10  20/01/2018   Female           suspected
#> 3  ciqafy     2018-01-03  13/01/2018   female            probable
#> 4  miqnfj     2018-01-09  19/01/2018   Female           suspected
#> 5  fikzdx     2018-01-07  17/01/2018     Male            PROBABLE
#> 6  llwgke     2018-01-11  21/01/2018   female          Not.a.Case
#> 7  wpekae     2018-01-05  15/01/2018   female           confirmed
#> 8  biahsq     2018-01-10  20/01/2018     Male          not a case
#> 9  zrdzhp     2018-01-04  14/01/2018   Female            probable
#> 10 gdurww     2018-01-08  18/01/2018     Male           Confirmed
#> 11 zhapbd     2018-01-03  13/01/2018   female           Confirmed
#> 12 afnvrv     2018-01-04  14/01/2018   Female           confirmed
#> 13 kqcbzo     2018-01-11  21/01/2018   female           Confirmed
#> 14 lnajcu     2018-01-05  15/01/2018     Male            probable
#> 15 umcose     2018-01-07  17/01/2018   FEMALE          not a case
#> 16 yzzpjy     2018-01-10  20/01/2018   female          Not.a.Case
#> 17 udajyn     2018-01-11  21/01/2018     male           Confirmed
#> 18 dzieiv     2018-01-03  13/01/2018   FEMALE         suspected  
#> 19 hkpxva     2018-01-11  21/01/2018     MALE           confirmed
#> 20 xxrain     2018-01-10  20/01/2018     male         suspected  
#>           messy/dates       lat      lon
#> 1  that's 24/12/1989! 11.851853 48.89684
#> 2          2018_10_17 10.650441 47.48114
#> 3          2018 10 19 13.107982 47.09722
#> 4          2018 10 19 15.430583 49.05794
#> 5          01-12-2001 12.116750 49.60121
#> 6     // 24//12//1989 14.491968 48.96044
#> 7                <NA> 13.425856 47.85849
#> 8              female 10.347720 47.01454
#> 9              female 16.337760 48.28741
#> 10               <NA> 12.747550 47.73664
#> 11         2018-10-18  8.846234 48.98780
#> 12         01-12-2001 16.345458 48.29515
#> 13    // 24//12//1989 13.020115 49.54814
#> 14         01-12-2001 15.843829 47.61313
#> 15         2018-10-18 13.535986 48.07265
#> 16             female 13.807798 47.61683
#> 17               <NA> 11.310754 47.93489
#> 18               male 13.833525 48.83712
#> 19    // 24//12//1989 10.071308 49.33979
#> 20         2018_10_17 12.038780 47.44533

## clean variable names, store in new object, show results
clean_data <- clean_data(toy_data, error_tolerance = 0.1)
clean_data

#>        id date_of_onset  discharge gender epi_case_definition       messy_dates
#> 1  fnkgjs    2018-01-11 2018-01-21 female           confirmed that_s_24_12_1989
#> 2  vbjzfx    2018-01-10 2018-01-20 female           suspected        2018_10_17
#> 3  ciqafy    2018-01-03 2018-01-13 female            probable        2018_10_19
#> 4  miqnfj    2018-01-09 2018-01-19 female           suspected        2018_10_19
#> 5  fikzdx    2018-01-07 2018-01-17   male            probable        01_12_2001
#> 6  llwgke    2018-01-11 2018-01-21 female          not_a_case        24_12_1989
#> 7  wpekae    2018-01-05 2018-01-15 female           confirmed              <NA>
#> 8  biahsq    2018-01-10 2018-01-20   male          not_a_case            female
#> 9  zrdzhp    2018-01-04 2018-01-14 female            probable            female
#> 10 gdurww    2018-01-08 2018-01-18   male           confirmed              <NA>
#> 11 zhapbd    2018-01-03 2018-01-13 female           confirmed        2018_10_18
#> 12 afnvrv    2018-01-04 2018-01-14 female           confirmed        01_12_2001
#> 13 kqcbzo    2018-01-11 2018-01-21 female           confirmed        24_12_1989
#> 14 lnajcu    2018-01-05 2018-01-15   male            probable        01_12_2001
#> 15 umcose    2018-01-07 2018-01-17 female          not_a_case        2018_10_18
#> 16 yzzpjy    2018-01-10 2018-01-20 female          not_a_case            female
#> 17 udajyn    2018-01-11 2018-01-21   male           confirmed              <NA>
#> 18 dzieiv    2018-01-03 2018-01-13 female           suspected              male
#> 19 hkpxva    2018-01-11 2018-01-21   male           confirmed        24_12_1989
#> 20 xxrain    2018-01-10 2018-01-20   male           suspected        2018_10_17
#>          lat      lon
#> 1  11.851853 48.89684
#> 2  10.650441 47.48114
#> 3  13.107982 47.09722
#> 4  15.430583 49.05794
#> 5  12.116750 49.60121
#> 6  14.491968 48.96044
#> 7  13.425856 47.85849
#> 8  10.347720 47.01454
#> 9  16.337760 48.28741
#> 10 12.747550 47.73664
#> 11  8.846234 48.98780
#> 12 16.345458 48.29515
#> 13 13.020115 49.54814
#> 14 15.843829 47.61313
#> 15 13.535986 48.07265
#> 16 13.807798 47.61683
#> 17 11.310754 47.93489
#> 18 13.833525 48.83712
#> 19 10.071308 49.33979
#> 20 12.038780 47.44533

clean_data2 <- clean_data(toy_data, error_tolerance = 0.8)
clean_data2

#>        id date_of_onset  discharge gender epi_case_definition messy_dates
#> 1  fnkgjs    2018-01-11 2018-01-21 female           confirmed  1989-12-24
#> 2  vbjzfx    2018-01-10 2018-01-20 female           suspected  2018-10-17
#> 3  ciqafy    2018-01-03 2018-01-13 female            probable  2018-10-19
#> 4  miqnfj    2018-01-09 2018-01-19 female           suspected  2018-10-19
#> 5  fikzdx    2018-01-07 2018-01-17   male            probable  2001-12-01
#> 6  llwgke    2018-01-11 2018-01-21 female          not_a_case  1989-12-24
#> 7  wpekae    2018-01-05 2018-01-15 female           confirmed        <NA>
#> 8  biahsq    2018-01-10 2018-01-20   male          not_a_case        <NA>
#> 9  zrdzhp    2018-01-04 2018-01-14 female            probable        <NA>
#> 10 gdurww    2018-01-08 2018-01-18   male           confirmed        <NA>
#> 11 zhapbd    2018-01-03 2018-01-13 female           confirmed  2018-10-18
#> 12 afnvrv    2018-01-04 2018-01-14 female           confirmed  2001-12-01
#> 13 kqcbzo    2018-01-11 2018-01-21 female           confirmed  1989-12-24
#> 14 lnajcu    2018-01-05 2018-01-15   male            probable  2001-12-01
#> 15 umcose    2018-01-07 2018-01-17 female          not_a_case  2018-10-18
#> 16 yzzpjy    2018-01-10 2018-01-20 female          not_a_case        <NA>
#> 17 udajyn    2018-01-11 2018-01-21   male           confirmed        <NA>
#> 18 dzieiv    2018-01-03 2018-01-13 female           suspected        <NA>
#> 19 hkpxva    2018-01-11 2018-01-21   male           confirmed  1989-12-24
#> 20 xxrain    2018-01-10 2018-01-20   male           suspected  2018-10-17
#>          lat      lon
#> 1  11.851853 48.89684
#> 2  10.650441 47.48114
#> 3  13.107982 47.09722
#> 4  15.430583 49.05794
#> 5  12.116750 49.60121
#> 6  14.491968 48.96044
#> 7  13.425856 47.85849
#> 8  10.347720 47.01454
#> 9  16.337760 48.28741
#> 10 12.747550 47.73664
#> 11  8.846234 48.98780
#> 12 16.345458 48.29515
#> 13 13.020115 49.54814
#> 14 15.843829 47.61313
#> 15 13.535986 48.07265
#> 16 13.807798 47.61683
#> 17 11.310754 47.93489
#> 18 13.833525 48.83712
#> 19 10.071308 49.33979
#> 20 12.038780 47.44533

## clean variable names, but keep our "messy/dates" column
to_protect <- names(toy_data) %in% "messy/dates"
clean_data3 <- clean_data(toy_data,
                          error_tolerance = 0.8,
                          protect = to_protect
                         )
clean_data3

#>        id date_of_onset  discharge gender epi_case_definition
#> 1  fnkgjs    2018-01-11 2018-01-21 female           confirmed
#> 2  vbjzfx    2018-01-10 2018-01-20 female           suspected
#> 3  ciqafy    2018-01-03 2018-01-13 female            probable
#> 4  miqnfj    2018-01-09 2018-01-19 female           suspected
#> 5  fikzdx    2018-01-07 2018-01-17   male            probable
#> 6  llwgke    2018-01-11 2018-01-21 female          not_a_case
#> 7  wpekae    2018-01-05 2018-01-15 female           confirmed
#> 8  biahsq    2018-01-10 2018-01-20   male          not_a_case
#> 9  zrdzhp    2018-01-04 2018-01-14 female            probable
#> 10 gdurww    2018-01-08 2018-01-18   male           confirmed
#> 11 zhapbd    2018-01-03 2018-01-13 female           confirmed
#> 12 afnvrv    2018-01-04 2018-01-14 female           confirmed
#> 13 kqcbzo    2018-01-11 2018-01-21 female           confirmed
#> 14 lnajcu    2018-01-05 2018-01-15   male            probable
#> 15 umcose    2018-01-07 2018-01-17 female          not_a_case
#> 16 yzzpjy    2018-01-10 2018-01-20 female          not_a_case
#> 17 udajyn    2018-01-11 2018-01-21   male           confirmed
#> 18 dzieiv    2018-01-03 2018-01-13 female           suspected
#> 19 hkpxva    2018-01-11 2018-01-21   male           confirmed
#> 20 xxrain    2018-01-10 2018-01-20   male           suspected
#>           messy/dates       lat      lon
#> 1  that's 24/12/1989! 11.851853 48.89684
#> 2          2018_10_17 10.650441 47.48114
#> 3          2018 10 19 13.107982 47.09722
#> 4          2018 10 19 15.430583 49.05794
#> 5          01-12-2001 12.116750 49.60121
#> 6     // 24//12//1989 14.491968 48.96044
#> 7                <NA> 13.425856 47.85849
#> 8              female 10.347720 47.01454
#> 9              female 16.337760 48.28741
#> 10               <NA> 12.747550 47.73664
#> 11         2018-10-18  8.846234 48.98780
#> 12         01-12-2001 16.345458 48.29515
#> 13    // 24//12//1989 13.020115 49.54814
#> 14         01-12-2001 15.843829 47.61313
#> 15         2018-10-18 13.535986 48.07265
#> 16             female 13.807798 47.61683
#> 17               <NA> 11.310754 47.93489
#> 18               male 13.833525 48.83712
#> 19    // 24//12//1989 10.071308 49.33979
#> 20         2018_10_17 12.038780 47.44533

## Using a wordlist  -------------------------------

# location data with mis-spellings, French, and English.
messy_locations <- c("hopsital", "h\u00f4pital", "hospital",
                     "m\u00e9dical", "clinic",
                     "feild", "field")
toy_data$location <- factor(sample(messy_locations, 20, replace = TRUE))

# show data 
toy_data$location

#>  [1] feild    hospital hôpital  médical  feild    hospital feild    clinic  
#>  [9] hospital clinic   field    field    clinic   feild    médical  médical 
#> [17] hopsital médical  hospital feild   
#> Levels: clinic feild field hôpital hopsital hospital médical

# add a wordlist
wordlist <- data.frame(
  from  = c("hopsital", "hopital",  "medical", "feild"),
  to    = c("hospital", "hospital", "clinic",  "field"),
  variables = rep("location", 4),
  stringsAsFactors = FALSE
)

clean_data4 <- clean_data(toy_data,
                          wordlists     = wordlist,
                          spelling_vars = "variables"
                         )
clean_data4

#>        id date_of_onset  discharge gender epi_case_definition messy_dates
#> 1  fnkgjs    2018-01-11 2018-01-21 female           confirmed  1989-12-24
#> 2  vbjzfx    2018-01-10 2018-01-20 female           suspected  2018-10-17
#> 3  ciqafy    2018-01-03 2018-01-13 female            probable  2018-10-19
#> 4  miqnfj    2018-01-09 2018-01-19 female           suspected  2018-10-19
#> 5  fikzdx    2018-01-07 2018-01-17   male            probable  2001-12-01
#> 6  llwgke    2018-01-11 2018-01-21 female          not_a_case  1989-12-24
#> 7  wpekae    2018-01-05 2018-01-15 female           confirmed        <NA>
#> 8  biahsq    2018-01-10 2018-01-20   male          not_a_case        <NA>
#> 9  zrdzhp    2018-01-04 2018-01-14 female            probable        <NA>
#> 10 gdurww    2018-01-08 2018-01-18   male           confirmed        <NA>
#> 11 zhapbd    2018-01-03 2018-01-13 female           confirmed  2018-10-18
#> 12 afnvrv    2018-01-04 2018-01-14 female           confirmed  2001-12-01
#> 13 kqcbzo    2018-01-11 2018-01-21 female           confirmed  1989-12-24
#> 14 lnajcu    2018-01-05 2018-01-15   male            probable  2001-12-01
#> 15 umcose    2018-01-07 2018-01-17 female          not_a_case  2018-10-18
#> 16 yzzpjy    2018-01-10 2018-01-20 female          not_a_case        <NA>
#> 17 udajyn    2018-01-11 2018-01-21   male           confirmed        <NA>
#> 18 dzieiv    2018-01-03 2018-01-13 female           suspected        <NA>
#> 19 hkpxva    2018-01-11 2018-01-21   male           confirmed  1989-12-24
#> 20 xxrain    2018-01-10 2018-01-20   male           suspected  2018-10-17
#>          lat      lon location
#> 1  11.851853 48.89684    field
#> 2  10.650441 47.48114 hospital
#> 3  13.107982 47.09722 hospital
#> 4  15.430583 49.05794   clinic
#> 5  12.116750 49.60121    field
#> 6  14.491968 48.96044 hospital
#> 7  13.425856 47.85849    field
#> 8  10.347720 47.01454   clinic
#> 9  16.337760 48.28741 hospital
#> 10 12.747550 47.73664   clinic
#> 11  8.846234 48.98780    field
#> 12 16.345458 48.29515    field
#> 13 13.020115 49.54814   clinic
#> 14 15.843829 47.61313    field
#> 15 13.535986 48.07265   clinic
#> 16 13.807798 47.61683   clinic
#> 17 11.310754 47.93489 hospital
#> 18 13.833525 48.83712   clinic
#> 19 10.071308 49.33979 hospital
#> 20 12.038780 47.44533    field

clean_data4$location

#>  [1] field    hospital hospital clinic   field    hospital field    clinic  
#>  [9] hospital clinic   field    field    clinic   field    clinic   clinic  
#> [17] hospital clinic   hospital field   
#> Levels: hospital clinic field