Clean a data.frame

This function applies several cleaning procedures to an input data.frame, by standardising variable names, labels used categorical variables (characters of factors), and setting dates to Date objects. Optionally, an intelligent date search can be used on character strings to extract dates from various formats mixed with other text. See details for more information.

clean_data(x, sep = "_", force_Date = TRUE, guess_dates = TRUE,
  error_tolerance = 0.5, quiet = FALSE)

Arguments

x
a data.frame

sep
The separator used between words, and defaults to the underscore _.

force_Date
a logical indicating if POSIXct and POSIXlt objects should be converted to Date objects; defaults to TRUE; you should use this if your dates are only precise to the day (i.e. no time information within days).

guess_dates
a logical indicating if dates should be guessed in columns storing character strings or factors; this feature is experimental; see guess_dates() for more information.

error_tolerance
a number between 0 and 1 indicating the proportion of entries which cannot be identified as dates to be tolerated; if this proportion is exceeded, the original vector is returned, and a message is issued; defaults to 0.1 (10 percent)

quiet
a logical indicating if messages should be displayed to the console (TRUE, default); set to FALSE to silence messages

Value

A data.frame with standardised labels for characters and factors.

Examples

## make toy data
onsets <- as.Date("2018-01-01") + sample(1:10, 20, replace = TRUE)
discharge <- format(as.Date(onsets) + 10, "%d/%m/%Y")
genders <- c("male", "female", "FEMALE", "Male", "Female", "MALE")
gender <- sample(genders, 20, replace = TRUE)
case_types <- c("confirmed", "probable", "suspected", "not a case",
                "Confirmed", "PROBABLE", "suspected  ", "Not.a.Case")
messy_dates <- sample(
                 c("01-12-2001", "male", "female", "2018-10-18", "2018_10_17",
                   "2018 10 19", "// 24//12//1989", NA, "that's 24/12/1989!"),
                 20, replace = TRUE)
case <- factor(sample(case_types, 20, replace = TRUE))
toy_data <- data.frame("Date of Onset." = onsets,
                       "DisCharge.." = discharge,
                       "GENDER_ " = gender,
                       "Épi.Case_définition" = case,
                       "messy/dates" = messy_dates)
## show data
toy_data

#>    Date.of.Onset. DisCharge.. GENDER_. Épi.Case_définition        messy.dates
#> 1      2018-01-02  12/01/2018   female         suspected      // 24//12//1989
#> 2      2018-01-10  20/01/2018   Female           suspected         2018_10_17
#> 3      2018-01-08  18/01/2018   Female           Confirmed         2018 10 19
#> 4      2018-01-03  13/01/2018   female           suspected         2018 10 19
#> 5      2018-01-02  12/01/2018     MALE           Confirmed         01-12-2001
#> 6      2018-01-06  16/01/2018   Female            probable    // 24//12//1989
#> 7      2018-01-06  16/01/2018     male          Not.a.Case    // 24//12//1989
#> 8      2018-01-04  14/01/2018     Male           Confirmed that's 24/12/1989!
#> 9      2018-01-09  19/01/2018   Female           Confirmed that's 24/12/1989!
#> 10     2018-01-09  19/01/2018   Female           suspected         2018-10-18
#> 11     2018-01-10  20/01/2018     male          not a case         2018_10_17
#> 12     2018-01-03  13/01/2018   female           suspected             female
#> 13     2018-01-02  12/01/2018   female           confirmed               male
#> 14     2018-01-05  15/01/2018     Male          not a case         2018_10_17
#> 15     2018-01-06  16/01/2018   FEMALE          not a case         2018_10_17
#> 16     2018-01-03  13/01/2018   FEMALE           confirmed               <NA>
#> 17     2018-01-06  16/01/2018   Female          not a case               male
#> 18     2018-01-02  12/01/2018     MALE           Confirmed    // 24//12//1989
#> 19     2018-01-05  15/01/2018   female         suspected           01-12-2001
#> 20     2018-01-11  21/01/2018   female          not a case         2018-10-18

## clean variable names, store in new object, show results
clean_data <- clean_data(toy_data, error_tolerance = 0.1)
clean_data

#>    date_of_onset  discharge gender epi_case_definition       messy_dates
#> 1     2018-01-02 2018-01-12 female           suspected        24_12_1989
#> 2     2018-01-10 2018-01-20 female           suspected        2018_10_17
#> 3     2018-01-08 2018-01-18 female           confirmed        2018_10_19
#> 4     2018-01-03 2018-01-13 female           suspected        2018_10_19
#> 5     2018-01-02 2018-01-12   male           confirmed        01_12_2001
#> 6     2018-01-06 2018-01-16 female            probable        24_12_1989
#> 7     2018-01-06 2018-01-16   male          not_a_case        24_12_1989
#> 8     2018-01-04 2018-01-14   male           confirmed that_s_24_12_1989
#> 9     2018-01-09 2018-01-19 female           confirmed that_s_24_12_1989
#> 10    2018-01-09 2018-01-19 female           suspected        2018_10_18
#> 11    2018-01-10 2018-01-20   male          not_a_case        2018_10_17
#> 12    2018-01-03 2018-01-13 female           suspected            female
#> 13    2018-01-02 2018-01-12 female           confirmed              male
#> 14    2018-01-05 2018-01-15   male          not_a_case        2018_10_17
#> 15    2018-01-06 2018-01-16 female          not_a_case        2018_10_17
#> 16    2018-01-03 2018-01-13 female           confirmed              <NA>
#> 17    2018-01-06 2018-01-16 female          not_a_case              male
#> 18    2018-01-02 2018-01-12   male           confirmed        24_12_1989
#> 19    2018-01-05 2018-01-15 female           suspected        01_12_2001
#> 20    2018-01-11 2018-01-21 female          not_a_case        2018_10_18

clean_data2 <- clean_data(toy_data, error_tolerance = 0.8)
clean_data2

#>    date_of_onset  discharge gender epi_case_definition messy_dates
#> 1     2018-01-02 2018-01-12 female           suspected  1989-12-24
#> 2     2018-01-10 2018-01-20 female           suspected  2018-10-17
#> 3     2018-01-08 2018-01-18 female           confirmed  2018-10-19
#> 4     2018-01-03 2018-01-13 female           suspected  2018-10-19
#> 5     2018-01-02 2018-01-12   male           confirmed  2001-12-01
#> 6     2018-01-06 2018-01-16 female            probable  1989-12-24
#> 7     2018-01-06 2018-01-16   male          not_a_case  1989-12-24
#> 8     2018-01-04 2018-01-14   male           confirmed  1989-12-24
#> 9     2018-01-09 2018-01-19 female           confirmed  1989-12-24
#> 10    2018-01-09 2018-01-19 female           suspected  2018-10-18
#> 11    2018-01-10 2018-01-20   male          not_a_case  2018-10-17
#> 12    2018-01-03 2018-01-13 female           suspected        <NA>
#> 13    2018-01-02 2018-01-12 female           confirmed        <NA>
#> 14    2018-01-05 2018-01-15   male          not_a_case  2018-10-17
#> 15    2018-01-06 2018-01-16 female          not_a_case  2018-10-17
#> 16    2018-01-03 2018-01-13 female           confirmed        <NA>
#> 17    2018-01-06 2018-01-16 female          not_a_case        <NA>
#> 18    2018-01-02 2018-01-12   male           confirmed  1989-12-24
#> 19    2018-01-05 2018-01-15 female           suspected  2001-12-01
#> 20    2018-01-11 2018-01-21 female          not_a_case  2018-10-18