Handle dates data

This function detects variables of data.frame which are effectively representing dates, and converts them to Date objects. When variables are character strings or factors, the function will try to convert dates with various pre-defined formats (see details). For each variable, the most common date format is automatically detected, and dates not following it are set to NA (i.e. missing). It uses a tolerance threshold for the amount of entries which cannot be converted to date (error_tolerance). By default, tolerance is set to 0.1, meaning 10 for a given variable. If there are more errors, this variable is assumed not to be a date, and left untouched.

clean_dates(x, force_Date = TRUE, guess_dates = TRUE,
  error_tolerance = 0.5, ..., classes = NULL)

Arguments

x
a data.frame

force_Date
a logical or integer vector indicating the columns . If logical, indicating if POSIXct and POSIXlt objects should be converted to Date objects; defaults to TRUE; you should use this if your dates are only precise to the day (i.e. no time information within days).

guess_dates
a logical or integer vector indicating which columns should be guessed , assuming these columns store character strings or factors; this feature is experimental; see guess_dates() for more information.

error_tolerance
a number between 0 and 1 indicating the proportion of entries which cannot be identified as dates to be tolerated; if this proportion is exceeded, the original vector is returned, and a message is issued; defaults to 0.1 (10 percent)

...
further arguments passed on to guess_dates()

classes
a vector of class definitions for each of the columns. If this is not provided, the classes will be read from the columns themselves. Practically, this is used in clean_data() to mark columns as protected.

Value

A data.frame with standardised dates.

See also

Examples

## make toy data
onsets <- as.POSIXct("2018-01-01", tz = "UTC")
onsets <- seq(onsets, by = "1 day", length.out = 10)
onsets <- sample(onsets, 20, replace = TRUE)
onsets2 <- format(as.Date(onsets), "%d/%m/%Y")
onsets3 <- format(as.Date(onsets), "%d %m %Y")
outcomes <- onsets + 1e7
admissions <- onsets + 86400 + sample(86400, 20)
admissions[1:5] <- NA
discharges <- admissions + (86400 * sample(5, 20, replace = TRUE)) + sample(86400, 20)
onset_with_errors <- onsets2
onset_with_errors[c(1,20)] <- c("male", "confirmed")
mixed_info <- onsets3
mixed_info[1:10] <- sample(c("bleeding", "fever"), 10, replace = TRUE)
gender <- sample(c("male", "female"), 20, replace = TRUE)
case_type <- c("confirmed", "probable", "suspected", "not a case")
case <- sample(case_type, 20, replace = TRUE)
toy_data <- data.frame("Date of Onset." = onsets,
                       "onset 2" = onsets2,
                       "ONSET 3" = onsets3,
                       "onset_4" = onset_with_errors,
                       "date admission" = admissions,
                       "DATE.of.DISCHARGE" = discharges,
                       "GENDER_ " = gender,
                       "Épi.Case_définition" = case,
                       "date of admission" = admissions,
                       "Date-of_discharge" = discharges,
                       "extra" = mixed_info,
                       stringsAsFactors = FALSE,
                       check.names = FALSE)
## show data
toy_data

#>    Date of Onset.    onset 2    ONSET 3    onset_4      date admission
#> 1      2018-01-03 03/01/2018 03 01 2018       male                <NA>
#> 2      2018-01-05 05/01/2018 05 01 2018 05/01/2018                <NA>
#> 3      2018-01-08 08/01/2018 08 01 2018 08/01/2018                <NA>
#> 4      2018-01-08 08/01/2018 08 01 2018 08/01/2018                <NA>
#> 5      2018-01-07 07/01/2018 07 01 2018 07/01/2018                <NA>
#> 6      2018-01-05 05/01/2018 05 01 2018 05/01/2018 2018-01-06 07:50:02
#> 7      2018-01-06 06/01/2018 06 01 2018 06/01/2018 2018-01-07 08:03:46
#> 8      2018-01-03 03/01/2018 03 01 2018 03/01/2018 2018-01-04 05:44:32
#> 9      2018-01-02 02/01/2018 02 01 2018 02/01/2018 2018-01-03 10:06:44
#> 10     2018-01-03 03/01/2018 03 01 2018 03/01/2018 2018-01-04 12:17:25
#> 11     2018-01-05 05/01/2018 05 01 2018 05/01/2018 2018-01-06 04:27:01
#> 12     2018-01-06 06/01/2018 06 01 2018 06/01/2018 2018-01-07 07:49:58
#> 13     2018-01-08 08/01/2018 08 01 2018 08/01/2018 2018-01-09 19:36:26
#> 14     2018-01-06 06/01/2018 06 01 2018 06/01/2018 2018-01-07 17:58:02
#> 15     2018-01-08 08/01/2018 08 01 2018 08/01/2018 2018-01-09 12:45:36
#> 16     2018-01-03 03/01/2018 03 01 2018 03/01/2018 2018-01-04 23:04:26
#> 17     2018-01-09 09/01/2018 09 01 2018 09/01/2018 2018-01-10 02:29:48
#> 18     2018-01-05 05/01/2018 05 01 2018 05/01/2018 2018-01-06 17:41:15
#> 19     2018-01-09 09/01/2018 09 01 2018 09/01/2018 2018-01-10 20:20:08
#> 20     2018-01-09 09/01/2018 09 01 2018  confirmed 2018-01-10 15:26:47
#>      DATE.of.DISCHARGE GENDER_  Épi.Case_définition   date of admission
#> 1                 <NA>     male           confirmed                <NA>
#> 2                 <NA>   female           suspected                <NA>
#> 3                 <NA>     male           suspected                <NA>
#> 4                 <NA>   female           confirmed                <NA>
#> 5                 <NA>   female            probable                <NA>
#> 6  2018-01-12 05:04:43   female          not a case 2018-01-06 07:50:02
#> 7  2018-01-11 22:37:10   female          not a case 2018-01-07 08:03:46
#> 8  2018-01-08 02:53:06   female            probable 2018-01-04 05:44:32
#> 9  2018-01-08 02:25:44     male            probable 2018-01-03 10:06:44
#> 10 2018-01-06 15:43:15     male           confirmed 2018-01-04 12:17:25
#> 11 2018-01-08 22:49:10     male           suspected 2018-01-06 04:27:01
#> 12 2018-01-12 22:52:42     male          not a case 2018-01-07 07:49:58
#> 13 2018-01-13 13:07:59     male            probable 2018-01-09 19:36:26
#> 14 2018-01-12 04:08:05   female           suspected 2018-01-07 17:58:02
#> 15 2018-01-10 19:11:57   female           confirmed 2018-01-09 12:45:36
#> 16 2018-01-07 17:31:41   female          not a case 2018-01-04 23:04:26
#> 17 2018-01-11 16:35:26   female          not a case 2018-01-10 02:29:48
#> 18 2018-01-10 15:14:56     male           suspected 2018-01-06 17:41:15
#> 19 2018-01-13 17:13:09   female           confirmed 2018-01-10 20:20:08
#> 20 2018-01-12 01:01:40   female           suspected 2018-01-10 15:26:47
#>      Date-of_discharge      extra
#> 1                 <NA>      fever
#> 2                 <NA>   bleeding
#> 3                 <NA>   bleeding
#> 4                 <NA>   bleeding
#> 5                 <NA>      fever
#> 6  2018-01-12 05:04:43   bleeding
#> 7  2018-01-11 22:37:10   bleeding
#> 8  2018-01-08 02:53:06   bleeding
#> 9  2018-01-08 02:25:44   bleeding
#> 10 2018-01-06 15:43:15   bleeding
#> 11 2018-01-08 22:49:10 05 01 2018
#> 12 2018-01-12 22:52:42 06 01 2018
#> 13 2018-01-13 13:07:59 08 01 2018
#> 14 2018-01-12 04:08:05 06 01 2018
#> 15 2018-01-10 19:11:57 08 01 2018
#> 16 2018-01-07 17:31:41 03 01 2018
#> 17 2018-01-11 16:35:26 09 01 2018
#> 18 2018-01-10 15:14:56 05 01 2018
#> 19 2018-01-13 17:13:09 09 01 2018
#> 20 2018-01-12 01:01:40 09 01 2018

str(toy_data)

#> 'data.frame':	20 obs. of  11 variables:
#>  $ Date of Onset.     : POSIXct, format: "2018-01-03" "2018-01-05" ...
#>  $ onset 2            : chr  "03/01/2018" "05/01/2018" "08/01/2018" "08/01/2018" ...
#>  $ ONSET 3            : chr  "03 01 2018" "05 01 2018" "08 01 2018" "08 01 2018" ...
#>  $ onset_4            : chr  "male" "05/01/2018" "08/01/2018" "08/01/2018" ...
#>  $ date admission     : POSIXct, format: NA NA ...
#>  $ DATE.of.DISCHARGE  : POSIXct, format: NA NA ...
#>  $ GENDER_            : chr  "male" "female" "male" "female" ...
#>  $ Épi.Case_définition: chr  "confirmed" "suspected" "suspected" "confirmed" ...
#>  $ date of admission  : POSIXct, format: NA NA ...
#>  $ Date-of_discharge  : POSIXct, format: NA NA ...
#>  $ extra              : chr  "fever" "bleeding" "bleeding" "bleeding" ...

## clean variable names, store in new object, show results
clean_data <- clean_variable_names(toy_data)
clean_data1 <- clean_dates(clean_data, first_date = "2018-01-01")
clean_data1

#>    date_of_onset    onset_2    onset_3    onset_4 date_admission
#> 1     2018-01-03 2018-01-03 2018-01-03       <NA>           <NA>
#> 2     2018-01-05 2018-01-05 2018-01-05 2018-01-05           <NA>
#> 3     2018-01-08 2018-01-08 2018-01-08 2018-01-08           <NA>
#> 4     2018-01-08 2018-01-08 2018-01-08 2018-01-08           <NA>
#> 5     2018-01-07 2018-01-07 2018-01-07 2018-01-07           <NA>
#> 6     2018-01-05 2018-01-05 2018-01-05 2018-01-05     2018-01-06
#> 7     2018-01-06 2018-01-06 2018-01-06 2018-01-06     2018-01-07
#> 8     2018-01-03 2018-01-03 2018-01-03 2018-01-03     2018-01-04
#> 9     2018-01-02 2018-01-02 2018-01-02 2018-01-02     2018-01-03
#> 10    2018-01-03 2018-01-03 2018-01-03 2018-01-03     2018-01-04
#> 11    2018-01-05 2018-01-05 2018-01-05 2018-01-05     2018-01-06
#> 12    2018-01-06 2018-01-06 2018-01-06 2018-01-06     2018-01-07
#> 13    2018-01-08 2018-01-08 2018-01-08 2018-01-08     2018-01-09
#> 14    2018-01-06 2018-01-06 2018-01-06 2018-01-06     2018-01-07
#> 15    2018-01-08 2018-01-08 2018-01-08 2018-01-08     2018-01-09
#> 16    2018-01-03 2018-01-03 2018-01-03 2018-01-03     2018-01-04
#> 17    2018-01-09 2018-01-09 2018-01-09 2018-01-09     2018-01-10
#> 18    2018-01-05 2018-01-05 2018-01-05 2018-01-05     2018-01-06
#> 19    2018-01-09 2018-01-09 2018-01-09 2018-01-09     2018-01-10
#> 20    2018-01-09 2018-01-09 2018-01-09       <NA>     2018-01-10
#>    date_of_discharge gender epi_case_definition date_of_admission
#> 1               <NA>   male           confirmed              <NA>
#> 2               <NA> female           suspected              <NA>
#> 3               <NA>   male           suspected              <NA>
#> 4               <NA> female           confirmed              <NA>
#> 5               <NA> female            probable              <NA>
#> 6         2018-01-12 female          not a case        2018-01-06
#> 7         2018-01-11 female          not a case        2018-01-07
#> 8         2018-01-08 female            probable        2018-01-04
#> 9         2018-01-08   male            probable        2018-01-03
#> 10        2018-01-06   male           confirmed        2018-01-04
#> 11        2018-01-08   male           suspected        2018-01-06
#> 12        2018-01-12   male          not a case        2018-01-07
#> 13        2018-01-13   male            probable        2018-01-09
#> 14        2018-01-12 female           suspected        2018-01-07
#> 15        2018-01-10 female           confirmed        2018-01-09
#> 16        2018-01-07 female          not a case        2018-01-04
#> 17        2018-01-11 female          not a case        2018-01-10
#> 18        2018-01-10   male           suspected        2018-01-06
#> 19        2018-01-13 female           confirmed        2018-01-10
#> 20        2018-01-12 female           suspected        2018-01-10
#>    date_of_discharge      extra
#> 1               <NA>       <NA>
#> 2               <NA>       <NA>
#> 3               <NA>       <NA>
#> 4               <NA>       <NA>
#> 5               <NA>       <NA>
#> 6         2018-01-12       <NA>
#> 7         2018-01-11       <NA>
#> 8         2018-01-08       <NA>
#> 9         2018-01-08       <NA>
#> 10        2018-01-06       <NA>
#> 11        2018-01-08 2018-01-05
#> 12        2018-01-12 2018-01-06
#> 13        2018-01-13 2018-01-08
#> 14        2018-01-12 2018-01-06
#> 15        2018-01-10 2018-01-08
#> 16        2018-01-07 2018-01-03
#> 17        2018-01-11 2018-01-09
#> 18        2018-01-10 2018-01-05
#> 19        2018-01-13 2018-01-09
#> 20        2018-01-12 2018-01-09

## Only clean the columns that have the words "date" or "admission" in them
the_date_cols <- grep("(date|admission)", names(clean_data))
the_date_cols

#> [1]  1  5  6  9 10

clean_data2 <- clean_dates(clean_data,
                           first_date  = "2018-01-01",
                           force_Date  = the_date_cols,
                           guess_dates = the_date_cols)
clean_data2

#>    date_of_onset    onset_2    onset_3    onset_4 date_admission
#> 1     2018-01-03 03/01/2018 03 01 2018       male           <NA>
#> 2     2018-01-05 05/01/2018 05 01 2018 05/01/2018           <NA>
#> 3     2018-01-08 08/01/2018 08 01 2018 08/01/2018           <NA>
#> 4     2018-01-08 08/01/2018 08 01 2018 08/01/2018           <NA>
#> 5     2018-01-07 07/01/2018 07 01 2018 07/01/2018           <NA>
#> 6     2018-01-05 05/01/2018 05 01 2018 05/01/2018     2018-01-06
#> 7     2018-01-06 06/01/2018 06 01 2018 06/01/2018     2018-01-07
#> 8     2018-01-03 03/01/2018 03 01 2018 03/01/2018     2018-01-04
#> 9     2018-01-02 02/01/2018 02 01 2018 02/01/2018     2018-01-03
#> 10    2018-01-03 03/01/2018 03 01 2018 03/01/2018     2018-01-04
#> 11    2018-01-05 05/01/2018 05 01 2018 05/01/2018     2018-01-06
#> 12    2018-01-06 06/01/2018 06 01 2018 06/01/2018     2018-01-07
#> 13    2018-01-08 08/01/2018 08 01 2018 08/01/2018     2018-01-09
#> 14    2018-01-06 06/01/2018 06 01 2018 06/01/2018     2018-01-07
#> 15    2018-01-08 08/01/2018 08 01 2018 08/01/2018     2018-01-09
#> 16    2018-01-03 03/01/2018 03 01 2018 03/01/2018     2018-01-04
#> 17    2018-01-09 09/01/2018 09 01 2018 09/01/2018     2018-01-10
#> 18    2018-01-05 05/01/2018 05 01 2018 05/01/2018     2018-01-06
#> 19    2018-01-09 09/01/2018 09 01 2018 09/01/2018     2018-01-10
#> 20    2018-01-09 09/01/2018 09 01 2018  confirmed     2018-01-10
#>    date_of_discharge gender epi_case_definition date_of_admission
#> 1               <NA>   male           confirmed              <NA>
#> 2               <NA> female           suspected              <NA>
#> 3               <NA>   male           suspected              <NA>
#> 4               <NA> female           confirmed              <NA>
#> 5               <NA> female            probable              <NA>
#> 6         2018-01-12 female          not a case        2018-01-06
#> 7         2018-01-11 female          not a case        2018-01-07
#> 8         2018-01-08 female            probable        2018-01-04
#> 9         2018-01-08   male            probable        2018-01-03
#> 10        2018-01-06   male           confirmed        2018-01-04
#> 11        2018-01-08   male           suspected        2018-01-06
#> 12        2018-01-12   male          not a case        2018-01-07
#> 13        2018-01-13   male            probable        2018-01-09
#> 14        2018-01-12 female           suspected        2018-01-07
#> 15        2018-01-10 female           confirmed        2018-01-09
#> 16        2018-01-07 female          not a case        2018-01-04
#> 17        2018-01-11 female          not a case        2018-01-10
#> 18        2018-01-10   male           suspected        2018-01-06
#> 19        2018-01-13 female           confirmed        2018-01-10
#> 20        2018-01-12 female           suspected        2018-01-10
#>    date_of_discharge      extra
#> 1               <NA>      fever
#> 2               <NA>   bleeding
#> 3               <NA>   bleeding
#> 4               <NA>   bleeding
#> 5               <NA>      fever
#> 6         2018-01-12   bleeding
#> 7         2018-01-11   bleeding
#> 8         2018-01-08   bleeding
#> 9         2018-01-08   bleeding
#> 10        2018-01-06   bleeding
#> 11        2018-01-08 05 01 2018
#> 12        2018-01-12 06 01 2018
#> 13        2018-01-13 08 01 2018
#> 14        2018-01-12 06 01 2018
#> 15        2018-01-10 08 01 2018
#> 16        2018-01-07 03 01 2018
#> 17        2018-01-11 09 01 2018
#> 18        2018-01-10 05 01 2018
#> 19        2018-01-13 09 01 2018
#> 20        2018-01-12 09 01 2018

str(clean_data2)

#> 'data.frame':	20 obs. of  11 variables:
#>  $ date_of_onset      : Date, format: "2018-01-03" "2018-01-05" ...
#>  $ onset_2            : chr  "03/01/2018" "05/01/2018" "08/01/2018" "08/01/2018" ...
#>  $ onset_3            : chr  "03 01 2018" "05 01 2018" "08 01 2018" "08 01 2018" ...
#>  $ onset_4            : chr  "male" "05/01/2018" "08/01/2018" "08/01/2018" ...
#>  $ date_admission     : Date, format: NA NA ...
#>  $ date_of_discharge  : Date, format: NA NA ...
#>  $ gender             : chr  "male" "female" "male" "female" ...
#>  $ epi_case_definition: chr  "confirmed" "suspected" "suspected" "confirmed" ...
#>  $ date_of_admission  : Date, format: NA NA ...
#>  $ date_of_discharge  : Date, format: NA NA ...
#>  $ extra              : chr  "fever" "bleeding" "bleeding" "bleeding" ...
#>  - attr(*, "comment")= Named chr  "Date of Onset." "onset 2" "ONSET 3" "onset_4" ...
#>   ..- attr(*, "names")= chr  "date_of_onset" "onset_2" "onset_3" "onset_4" ...

## A more complex example: clean date and admissions, but avoid the discharge
## column, since the timestamp is important
the_date_cols <- grepl("(date|admission)", names(clean_data))
discharge     <- grepl("discharge", names(clean_data))

## set names so that these are easier to track
names(the_date_cols) <- names(clean_data) -> names(discharge)

the_date_cols # columns we want

#>       date_of_onset             onset_2             onset_3             onset_4 
#>                TRUE               FALSE               FALSE               FALSE 
#>      date_admission   date_of_discharge              gender epi_case_definition 
#>                TRUE                TRUE               FALSE               FALSE 
#>   date_of_admission   date_of_discharge               extra 
#>                TRUE                TRUE               FALSE 

!discharge    # columns that are not the discharge columns ("!" means "not")

#>       date_of_onset             onset_2             onset_3             onset_4 
#>                TRUE                TRUE                TRUE                TRUE 
#>      date_admission   date_of_discharge              gender epi_case_definition 
#>                TRUE               FALSE                TRUE                TRUE 
#>   date_of_admission   date_of_discharge               extra 
#>                TRUE               FALSE                TRUE 

to_keep     <- the_date_cols & !discharge # removing the discharge column
clean_data3 <- clean_dates(clean_data,
                           first_date  = "2018-01-01",
                           force_Date  = to_keep,
                           guess_dates = to_keep)
clean_data3

#>    date_of_onset    onset_2    onset_3    onset_4 date_admission
#> 1     2018-01-03 03/01/2018 03 01 2018       male           <NA>
#> 2     2018-01-05 05/01/2018 05 01 2018 05/01/2018           <NA>
#> 3     2018-01-08 08/01/2018 08 01 2018 08/01/2018           <NA>
#> 4     2018-01-08 08/01/2018 08 01 2018 08/01/2018           <NA>
#> 5     2018-01-07 07/01/2018 07 01 2018 07/01/2018           <NA>
#> 6     2018-01-05 05/01/2018 05 01 2018 05/01/2018     2018-01-06
#> 7     2018-01-06 06/01/2018 06 01 2018 06/01/2018     2018-01-07
#> 8     2018-01-03 03/01/2018 03 01 2018 03/01/2018     2018-01-04
#> 9     2018-01-02 02/01/2018 02 01 2018 02/01/2018     2018-01-03
#> 10    2018-01-03 03/01/2018 03 01 2018 03/01/2018     2018-01-04
#> 11    2018-01-05 05/01/2018 05 01 2018 05/01/2018     2018-01-06
#> 12    2018-01-06 06/01/2018 06 01 2018 06/01/2018     2018-01-07
#> 13    2018-01-08 08/01/2018 08 01 2018 08/01/2018     2018-01-09
#> 14    2018-01-06 06/01/2018 06 01 2018 06/01/2018     2018-01-07
#> 15    2018-01-08 08/01/2018 08 01 2018 08/01/2018     2018-01-09
#> 16    2018-01-03 03/01/2018 03 01 2018 03/01/2018     2018-01-04
#> 17    2018-01-09 09/01/2018 09 01 2018 09/01/2018     2018-01-10
#> 18    2018-01-05 05/01/2018 05 01 2018 05/01/2018     2018-01-06
#> 19    2018-01-09 09/01/2018 09 01 2018 09/01/2018     2018-01-10
#> 20    2018-01-09 09/01/2018 09 01 2018  confirmed     2018-01-10
#>      date_of_discharge gender epi_case_definition date_of_admission
#> 1                 <NA>   male           confirmed              <NA>
#> 2                 <NA> female           suspected              <NA>
#> 3                 <NA>   male           suspected              <NA>
#> 4                 <NA> female           confirmed              <NA>
#> 5                 <NA> female            probable              <NA>
#> 6  2018-01-12 05:04:43 female          not a case        2018-01-06
#> 7  2018-01-11 22:37:10 female          not a case        2018-01-07
#> 8  2018-01-08 02:53:06 female            probable        2018-01-04
#> 9  2018-01-08 02:25:44   male            probable        2018-01-03
#> 10 2018-01-06 15:43:15   male           confirmed        2018-01-04
#> 11 2018-01-08 22:49:10   male           suspected        2018-01-06
#> 12 2018-01-12 22:52:42   male          not a case        2018-01-07
#> 13 2018-01-13 13:07:59   male            probable        2018-01-09
#> 14 2018-01-12 04:08:05 female           suspected        2018-01-07
#> 15 2018-01-10 19:11:57 female           confirmed        2018-01-09
#> 16 2018-01-07 17:31:41 female          not a case        2018-01-04
#> 17 2018-01-11 16:35:26 female          not a case        2018-01-10
#> 18 2018-01-10 15:14:56   male           suspected        2018-01-06
#> 19 2018-01-13 17:13:09 female           confirmed        2018-01-10
#> 20 2018-01-12 01:01:40 female           suspected        2018-01-10
#>      date_of_discharge      extra
#> 1                 <NA>      fever
#> 2                 <NA>   bleeding
#> 3                 <NA>   bleeding
#> 4                 <NA>   bleeding
#> 5                 <NA>      fever
#> 6  2018-01-12 05:04:43   bleeding
#> 7  2018-01-11 22:37:10   bleeding
#> 8  2018-01-08 02:53:06   bleeding
#> 9  2018-01-08 02:25:44   bleeding
#> 10 2018-01-06 15:43:15   bleeding
#> 11 2018-01-08 22:49:10 05 01 2018
#> 12 2018-01-12 22:52:42 06 01 2018
#> 13 2018-01-13 13:07:59 08 01 2018
#> 14 2018-01-12 04:08:05 06 01 2018
#> 15 2018-01-10 19:11:57 08 01 2018
#> 16 2018-01-07 17:31:41 03 01 2018
#> 17 2018-01-11 16:35:26 09 01 2018
#> 18 2018-01-10 15:14:56 05 01 2018
#> 19 2018-01-13 17:13:09 09 01 2018
#> 20 2018-01-12 01:01:40 09 01 2018

str(clean_data3)

#> 'data.frame':	20 obs. of  11 variables:
#>  $ date_of_onset      : Date, format: "2018-01-03" "2018-01-05" ...
#>  $ onset_2            : chr  "03/01/2018" "05/01/2018" "08/01/2018" "08/01/2018" ...
#>  $ onset_3            : chr  "03 01 2018" "05 01 2018" "08 01 2018" "08 01 2018" ...
#>  $ onset_4            : chr  "male" "05/01/2018" "08/01/2018" "08/01/2018" ...
#>  $ date_admission     : Date, format: NA NA ...
#>  $ date_of_discharge  : POSIXct, format: NA NA ...
#>  $ gender             : chr  "male" "female" "male" "female" ...
#>  $ epi_case_definition: chr  "confirmed" "suspected" "suspected" "confirmed" ...
#>  $ date_of_admission  : Date, format: NA NA ...
#>  $ date_of_discharge  : POSIXct, format: NA NA ...
#>  $ extra              : chr  "fever" "bleeding" "bleeding" "bleeding" ...
#>  - attr(*, "comment")= Named chr  "Date of Onset." "onset 2" "ONSET 3" "onset_4" ...
#>   ..- attr(*, "names")= chr  "date_of_onset" "onset_2" "onset_3" "onset_4" ...