Clean variable labels

This function standardises the labels of all characters of factors in a data.frame. It uses the standardisation implemented by epitrix::clean_labels() in the epitrix package. See ?epitrix::clean_labels for more information.

clean_variable_labels(x, classes = NULL, ...)

Arguments

x
a data.frame

classes
a vector of class definitions for each of the columns. If this is not provided, the classes will be read from the columns themselves. Practically, this is used in clean_data() to mark columns as protected.

...
further arguments passed to epitrix::clean_labels(); the most important is sep, which refers to the separator used between words, and defaults to the underscore _.

Value

A data.frame with standardised labels for characters and factors.

Examples

## make toy data
toy_data <- clean_variable_names(messy_data())

## clean variable labels, store in new object, show results
clean_data <- clean_variable_labels(toy_data)
clean_data

#>        id date_of_onset  discharge gender epi_case_definition messy_dates
#> 1  tlrazf    2018-01-02 12_01_2018 female           suspected  2018_10_18
#> 2  htdceq    2018-01-03 13_01_2018   male           confirmed  2018_10_19
#> 3  rufupp    2018-01-04 14_01_2018   male            probable  2018_10_17
#> 4  cfaybd    2018-01-10 20_01_2018 female           confirmed  2018_10_19
#> 5  imeblj    2018-01-04 14_01_2018   male           suspected      female
#> 6  xafrtq    2018-01-11 21_01_2018   male            probable  24_12_1989
#> 7  cpwtcv    2018-01-09 19_01_2018   male          not_a_case      female
#> 8  wstdfi    2018-01-10 20_01_2018 female            probable  2018_10_18
#> 9  dwmmui    2018-01-08 18_01_2018   male            probable  2018_10_17
#> 10 iivhlk    2018-01-06 16_01_2018 female            probable  24_12_1989
#> 11 dreypi    2018-01-11 21_01_2018 female           suspected      female
#> 12 vxdodq    2018-01-03 13_01_2018 female            probable  24_12_1989
#> 13 bhidue    2018-01-03 13_01_2018   male           confirmed  2018_10_19
#> 14 umlosx    2018-01-08 18_01_2018   male            probable  01_12_2001
#> 15 fhvbng    2018-01-06 16_01_2018 female            probable  24_12_1989
#> 16 pzgwcu    2018-01-05 15_01_2018 female           confirmed  01_12_2001
#> 17 rqfwlh    2018-01-05 15_01_2018 female          not_a_case  01_12_2001
#> 18 qgkqom    2018-01-10 20_01_2018   male            probable  24_12_1989
#> 19 tyhsdb    2018-01-04 14_01_2018   male           confirmed  2018_10_19
#> 20 sdtecy    2018-01-08 18_01_2018 female            probable  24_12_1989
#>          lat      lon
#> 1  13.841749 47.27340
#> 2  12.284943 48.68188
#> 3  11.706277 47.77016
#> 4  12.899716 46.48940
#> 5  13.833886 47.41627
#> 6  11.734825 45.97708
#> 7  15.300293 48.40350
#> 8  12.529048 48.55002
#> 9   9.713785 48.02836
#> 10  9.993236 48.89317
#> 11  8.898830 47.62344
#> 12 11.493604 48.60588
#> 13 12.731716 47.99513
#> 14 15.011566 47.47920
#> 15 17.334374 47.36098
#> 16 17.645113 47.36411
#> 17 10.959153 48.10659
#> 18 13.097629 49.17691
#> 19 11.456223 48.44739
#> 20 11.429530 50.27295