Clean variable labels

This function standardises the labels of all characters of factors in a data.frame. It uses the standardisation implemented by epitrix::clean_labels() in the epitrix package. See ?epitrix::clean_labels for more information.

clean_variable_labels(x, ...)

Arguments

x
a data.frame

...
further arguments passed to epitrix::clean_labels(); the most important is sep, which refers to the separator used between words, and defaults to the underscore _.

Value

A data.frame with standardised labels for characters and factors.

Examples

## make toy data
onsets <- as.Date("2018-01-01") + sample(1:10, 20, replace = TRUE)
genders <- c("male", "female", "FEMALE", "Male", "Female", "MALE")
gender <- sample(genders, 20, replace = TRUE)
case_types <- c("confirmed", "probable", "suspected", "not a case",
                "Confirmed", "PROBABLE", "suspected  ", "Not.a.Case")
case <- factor(sample(case_types, 20, replace = TRUE))
toy_data <- data.frame("Date of Onset." = onsets,
                       "GENDER_ " = gender,
                       "Épi.Case_définition" = case)
## show data
toy_data

#>    Date.of.Onset. GENDER_. Épi.Case_définition
#> 1      2018-01-09   female           suspected
#> 2      2018-01-09   FEMALE           confirmed
#> 3      2018-01-04     male           confirmed
#> 4      2018-01-09   Female         suspected  
#> 5      2018-01-02   female            probable
#> 6      2018-01-06   Female          not a case
#> 7      2018-01-07     Male          Not.a.Case
#> 8      2018-01-06     MALE            PROBABLE
#> 9      2018-01-10     male            PROBABLE
#> 10     2018-01-11     Male           suspected
#> 11     2018-01-08     male          Not.a.Case
#> 12     2018-01-02   FEMALE           confirmed
#> 13     2018-01-03   Female            PROBABLE
#> 14     2018-01-06   female          Not.a.Case
#> 15     2018-01-09     male            probable
#> 16     2018-01-07     male            probable
#> 17     2018-01-10   FEMALE           confirmed
#> 18     2018-01-04     Male           Confirmed
#> 19     2018-01-04   Female          not a case
#> 20     2018-01-08   Female          Not.a.Case

## clean variable names, store in new object, show results
clean_data <- clean_variable_names(toy_data)
clean_data

#>    date_of_onset gender epi_case_definition
#> 1     2018-01-09 female           suspected
#> 2     2018-01-09 FEMALE           confirmed
#> 3     2018-01-04   male           confirmed
#> 4     2018-01-09 Female         suspected  
#> 5     2018-01-02 female            probable
#> 6     2018-01-06 Female          not a case
#> 7     2018-01-07   Male          Not.a.Case
#> 8     2018-01-06   MALE            PROBABLE
#> 9     2018-01-10   male            PROBABLE
#> 10    2018-01-11   Male           suspected
#> 11    2018-01-08   male          Not.a.Case
#> 12    2018-01-02 FEMALE           confirmed
#> 13    2018-01-03 Female            PROBABLE
#> 14    2018-01-06 female          Not.a.Case
#> 15    2018-01-09   male            probable
#> 16    2018-01-07   male            probable
#> 17    2018-01-10 FEMALE           confirmed
#> 18    2018-01-04   Male           Confirmed
#> 19    2018-01-04 Female          not a case
#> 20    2018-01-08 Female          Not.a.Case