Check and clean spelling or codes of multiple variables in a data frame

This function allows you to clean your data according to pre-defined rules encapsulated in either a data frame or list of data frames. It has application for addressing mis-spellings and recoding variables (e.g. from electronic survey data).

By default, this applies the function clean_spelling() to all columns specified by the column names listed in spelling_vars, or, if a global dictionary is used, this includes all character and factor columns as well.

Global wordlists
A global wordlist is a set of definitions applied to all valid columns of x indiscriminantly. When a global dictioary is used, you might see several warnings appear because some variables were not found in the data.

.global spelling_var: If you want to apply a set of definitions to all valid columns in addition to specified columns, then you can include a .global group in the spelling_var column of your wordlists data frame. NOTE: specific variable definitions will override global defintions. For example: if you have a column for cardinal directions and a definiton for N = North, then the global variable N = no will not override that. See Example.

spelling_var = NULL: If you want your data frame to be applied to all character/factor columns indiscriminantly, then setting spelling_var = NULL will use that wordlist globally.

clean_variable_spelling(x = data.frame(), wordlists = list(),
  spelling_vars = 3, sort_by = NULL, classes = NULL)

Arguments

x
a data.frame

wordlists
a data frame or named list of data frames with at least two columns defining the word list to be used. If this is a data frame, a third column must be present to split the wordlists by column in x (see spelling_vars).

spelling_vars
character or integer. If wordlists is a data frame, then this column in defines the columns in x corresponding to each section of the wordlists data frame. This defaults to 3, indicating the third column is to be used.

sort_by
a character the column to be used for sorting the values in each data frame. If the incoming variables are factors, this determines how the resulting factors will be sorted.

classes
a vector of class definitions for each of the columns. If this is not provided, the classes will be read from the columns themselves. Practically, this is used in clean_data() to mark columns as protected.

Value

a data frame with re-defined data based on the dictionary

Note

This function will only parse character and factor columns to protect numeric and Date columns from conversion to character.

See also

Examples

# Set up wordlist ------------------------------------------------ 

yesno  <- c("Y", "N", "U", NA)
dyesno <- c("Yes", "No", "Unknown", "Missing")

treatment_administered  <- c(0:1, NA)
dtreatment_administered <- c("Yes", "No", "Missing")

facility  <- c(1:10, ".default") # define a .default key
dfacility <- c(sprintf("Facility %s", format(1:10)), "Unknown")

age_group  <- c(0, 10, 20, 30, 40, 50)
dage_group <- c("0-9", "10-19", "20-29", "30-39", "40-49", "50+")

wordlist <- data.frame(
  options = c(yesno, treatment_administered, facility, age_group),
  values  = c(dyesno, dtreatment_administered, dfacility, dage_group),
  grp = rep(c("readmission", "treatment_administered", "facility", "age_group"),
            c(4, 3, 11, 6)),
  orders  = c(1:4, 1:3, 1:11, 1:6),
  stringsAsFactors = FALSE
)

# Assigning global values ----------------------------------------

global_words <- data.frame(
  options = c("Y", "N", "U", "unk", "oui", NA),
  values  = c("yes", "no", "unknown", "unknown", "yes", "missing"),
  grp     = rep(".global", 6),
  orders  = rep(Inf, 6),
  stringsAsFactors = FALSE
)

wordlist <- rbind(wordlist, global_words, stringsAsFactors = FALSE)

# Generate example data ------------------------------------------
dat <- data.frame(
  # these have been defined
  readmission = sample(yesno, 50, replace = TRUE),
  treatment_administered = sample(treatment_administered, 50, replace = TRUE),
  facility = sample(c(facility[-11], LETTERS[1:3]), 50, replace = TRUE),
  age_group = sample(age_group, 50, replace = TRUE),
  # global values will catch these
  has_symptoms = sample(c(yesno, "unk", "oui"), 50, replace = TRUE),
  followup = sample(c(yesno, "unk", "oui"), 50, replace = TRUE),
  stringsAsFactors = FALSE
)

# Clean spelling based on wordlist ------------------------------ 

wordlist # show the wordlist

#>     options      values                    grp orders
#> 1         Y         Yes            readmission      1
#> 2         N          No            readmission      2
#> 3         U     Unknown            readmission      3
#> 4      <NA>     Missing            readmission      4
#> 5         0         Yes treatment_administered      1
#> 6         1          No treatment_administered      2
#> 7      <NA>     Missing treatment_administered      3
#> 8         1 Facility  1               facility      1
#> 9         2 Facility  2               facility      2
#> 10        3 Facility  3               facility      3
#> 11        4 Facility  4               facility      4
#> 12        5 Facility  5               facility      5
#> 13        6 Facility  6               facility      6
#> 14        7 Facility  7               facility      7
#> 15        8 Facility  8               facility      8
#> 16        9 Facility  9               facility      9
#> 17       10 Facility 10               facility     10
#> 18 .default     Unknown               facility     11
#> 19        0         0-9              age_group      1
#> 20       10       10-19              age_group      2
#> 21       20       20-29              age_group      3
#> 22       30       30-39              age_group      4
#> 23       40       40-49              age_group      5
#> 24       50         50+              age_group      6
#> 25        Y         yes                .global    Inf
#> 26        N          no                .global    Inf
#> 27        U     unknown                .global    Inf
#> 28      unk     unknown                .global    Inf
#> 29      oui         yes                .global    Inf
#> 30     <NA>     missing                .global    Inf

head(dat) # show the data

#>   readmission treatment_administered facility age_group has_symptoms followup
#> 1           N                     NA        B        30            N        U
#> 2           N                     NA        2        50            U     <NA>
#> 3           Y                     NA        B         0          unk        N
#> 4           N                      1        4        20            N        Y
#> 5           N                     NA       10         0            Y        N
#> 6           U                      1        C        50            Y      unk

head(clean_variable_spelling(dat, wordlists = wordlist, spelling_vars = "grp"))

#>   readmission treatment_administered    facility age_group has_symptoms
#> 1          No                Missing     Unknown     30-39           no
#> 2          No                Missing Facility  2       50+      unknown
#> 3         Yes                Missing     Unknown       0-9      unknown
#> 4          No                     No Facility  4     20-29           no
#> 5          No                Missing Facility 10       0-9          yes
#> 6     Unknown                     No     Unknown       50+          yes
#>   followup
#> 1  unknown
#> 2  missing
#> 3       no
#> 4      yes
#> 5       no
#> 6  unknown

# You can ensure the order of the factors are correct by specifying 
# a column that defines order.

dat[] <- lapply(dat, as.factor)
as.list(head(dat))

#> $readmission
#> [1] N N Y N N U
#> Levels: N U Y
#> 
#> $treatment_administered
#> [1] <NA> <NA> <NA> 1    <NA> 1   
#> Levels: 0 1
#> 
#> $facility
#> [1] B  2  B  4  10 C 
#> Levels: 1 10 2 3 4 5 6 7 8 9 A B C
#> 
#> $age_group
#> [1] 30 50 0  20 0  50
#> Levels: 0 10 20 30 40 50
#> 
#> $has_symptoms
#> [1] N   U   unk N   Y   Y  
#> Levels: N oui U unk Y
#> 
#> $followup
#> [1] U    <NA> N    Y    N    unk 
#> Levels: N oui U unk Y
#> 

res <- clean_variable_spelling(dat,
                               wordlists = wordlist,
                               spelling_vars = "grp",
                               sort_by = "orders")

#> Warning: Unknown levels in `f`: yes, no, unknown, missing

#> Warning: Unknown levels in `f`: yes, no, unknown, missing

#> Warning: Unknown levels in `f`: unknown, yes

#> Warning: Unknown levels in `f`: yes, no, unknown

head(res)

#>   readmission treatment_administered    facility age_group has_symptoms
#> 1          No                Missing     Unknown     30-39           no
#> 2          No                Missing Facility  2       50+      unknown
#> 3         Yes                Missing     Unknown       0-9      unknown
#> 4          No                     No Facility  4     20-29           no
#> 5          No                Missing Facility 10       0-9          yes
#> 6     Unknown                     No     Unknown       50+          yes
#>   followup
#> 1  unknown
#> 2  missing
#> 3       no
#> 4      yes
#> 5       no
#> 6  unknown

as.list(head(res))

#> $readmission
#> [1] No      No      Yes     No      No      Unknown
#> Levels: Yes No Unknown Missing
#> 
#> $treatment_administered
#> [1] Missing Missing Missing No      Missing No     
#> Levels: Yes No Missing
#> 
#> $facility
#> [1] Unknown     Facility  2 Unknown     Facility  4 Facility 10 Unknown    
#> 11 Levels: Facility  1 Facility  2 Facility  3 Facility  4 ... Unknown
#> 
#> $age_group
#> [1] 30-39 50+   0-9   20-29 0-9   50+  
#> Levels: 0-9 10-19 20-29 30-39 40-49 50+
#> 
#> $has_symptoms
#> [1] no      unknown unknown no      yes     yes    
#> Levels: yes no unknown missing
#> 
#> $followup
#> [1] unknown missing no      yes     no      unknown
#> Levels: yes no unknown missing
#>