Clean variable labels and fix spelling according to a wordlist

Clean variable labels and fix spelling according to a wordlist

clean_variables(x, sep = "_", wordlists = NULL, spelling_vars = 3,
  sort_by = NULL, protect = FALSE, classes = NULL,
  warn_spelling = FALSE)

Arguments

x
a data.frame

sep
The separator used between words, and defaults to the underscore _.

wordlists
a data frame or named list of data frames with at least two columns defining the word list to be used. If this is a data frame, a third column must be present to split the wordlists by column in x (see spelling_vars).

spelling_vars
character or integer. If wordlists is a data frame, then this column in defines the columns in x corresponding to each section of the wordlists data frame. This defaults to 3, indicating the third column is to be used.

sort_by
a character the column to be used for sorting the values in each data frame. If the incoming variables are factors, this determines how the resulting factors will be sorted.

protect
a logical or numeric vector defining the columns to protect from any manipulation. Note: columns in protect will override any columns in either force_Date or guess_dates.

classes
a vector of class definitions for each of the columns. If this is not provided, the classes will be read from the columns themselves. Practically, this is used in clean_data() to mark columns as protected.

warn_spelling
if TRUE, errors and warnings from clean_spelling() will be aggregated and presented for each column that issues them. The default value is FALSE, which means that all errors and warnings will be ignored.

See also

Examples

## make toy data
toy_data <- messy_data(20)

# location data with mis-spellings, French, and English.
messy_locations <- c("hopsital", "h\u00f4pital", "hospital",
                     "m\u00e9dical", "clinic",
                     "feild", "field")
toy_data$location <- sample(messy_locations, 20, replace = TRUE)

## show data
toy_data

#>       'ID Date of Onset. DisCharge.. GENDER_  Épi.Case_définition
#> 1  dewyhn     2018-01-06  16/01/2018   FEMALE           confirmed
#> 2  cgkhox     2018-01-10  20/01/2018   female           Confirmed
#> 3  qjvjib     2018-01-09  19/01/2018     MALE            probable
#> 4  oveifu     2018-01-04  14/01/2018     MALE          Not.a.Case
#> 5  aticgv     2018-01-03  13/01/2018   FEMALE         suspected  
#> 6  pycbjk     2018-01-08  18/01/2018   FEMALE           confirmed
#> 7  sgapix     2018-01-02  12/01/2018   FEMALE           suspected
#> 8  fcnmip     2018-01-04  14/01/2018     male           confirmed
#> 9  fridsl     2018-01-07  17/01/2018     Male           suspected
#> 10 bmqiyf     2018-01-02  12/01/2018   FEMALE          not a case
#> 11 vtptbs     2018-01-06  16/01/2018     male           suspected
#> 12 mhowzc     2018-01-02  12/01/2018   FEMALE          Not.a.Case
#> 13 ohoyig     2018-01-07  17/01/2018   female         suspected  
#> 14 miylys     2018-01-10  20/01/2018   female          Not.a.Case
#> 15 bpktsa     2018-01-08  18/01/2018     Male           Confirmed
#> 16 pwojeh     2018-01-10  20/01/2018   female           Confirmed
#> 17 hjaaws     2018-01-11  21/01/2018     Male           suspected
#> 18 iwqrju     2018-01-09  19/01/2018   female            PROBABLE
#> 19 fulhnw     2018-01-03  13/01/2018     Male            probable
#> 20 mbnyus     2018-01-09  19/01/2018   Female          not a case
#>           messy/dates      lat      lon location
#> 1          2018-10-18 11.77449 48.63144  hôpital
#> 2  that's 24/12/1989! 13.32132 47.51660  hôpital
#> 3                male 11.67528 46.47367  médical
#> 4  that's 24/12/1989! 12.33030 48.61684    feild
#> 5          2018-10-18 14.24602 49.22929  hôpital
#> 6              female 15.05639 48.15745    field
#> 7  that's 24/12/1989! 10.73084 49.40634  hôpital
#> 8          2018-10-18 14.83382 48.39680    feild
#> 9                male 15.42775 48.32642 hospital
#> 10         01-12-2001 11.62814 46.54066 hospital
#> 11         2018-10-18 10.81086 47.20070    field
#> 12               <NA> 13.72975 48.47001    feild
#> 13               <NA> 10.98397 46.84982 hospital
#> 14 that's 24/12/1989! 14.10999 47.72840   clinic
#> 15               <NA> 13.87213 48.45742   clinic
#> 16         2018_10_17 13.21075 47.98304    feild
#> 17             female 12.49252 47.45841   clinic
#> 18         2018_10_17 11.04702 48.87581    feild
#> 19 that's 24/12/1989! 13.78844 48.74084    field
#> 20         2018-10-18 15.02981 47.83519    field

# clean labels
clean_variables(toy_data) # by default, it's the same as clean_variable_lables

#>       'ID Date of Onset. DisCharge.. GENDER_  Épi.Case_définition
#> 1  dewyhn     2018-01-06  16_01_2018   female           confirmed
#> 2  cgkhox     2018-01-10  20_01_2018   female           confirmed
#> 3  qjvjib     2018-01-09  19_01_2018     male            probable
#> 4  oveifu     2018-01-04  14_01_2018     male          not_a_case
#> 5  aticgv     2018-01-03  13_01_2018   female           suspected
#> 6  pycbjk     2018-01-08  18_01_2018   female           confirmed
#> 7  sgapix     2018-01-02  12_01_2018   female           suspected
#> 8  fcnmip     2018-01-04  14_01_2018     male           confirmed
#> 9  fridsl     2018-01-07  17_01_2018     male           suspected
#> 10 bmqiyf     2018-01-02  12_01_2018   female          not_a_case
#> 11 vtptbs     2018-01-06  16_01_2018     male           suspected
#> 12 mhowzc     2018-01-02  12_01_2018   female          not_a_case
#> 13 ohoyig     2018-01-07  17_01_2018   female           suspected
#> 14 miylys     2018-01-10  20_01_2018   female          not_a_case
#> 15 bpktsa     2018-01-08  18_01_2018     male           confirmed
#> 16 pwojeh     2018-01-10  20_01_2018   female           confirmed
#> 17 hjaaws     2018-01-11  21_01_2018     male           suspected
#> 18 iwqrju     2018-01-09  19_01_2018   female            probable
#> 19 fulhnw     2018-01-03  13_01_2018     male            probable
#> 20 mbnyus     2018-01-09  19_01_2018   female          not_a_case
#>          messy/dates      lat      lon location
#> 1         2018_10_18 11.77449 48.63144  hopital
#> 2  that_s_24_12_1989 13.32132 47.51660  hopital
#> 3               male 11.67528 46.47367  medical
#> 4  that_s_24_12_1989 12.33030 48.61684    feild
#> 5         2018_10_18 14.24602 49.22929  hopital
#> 6             female 15.05639 48.15745    field
#> 7  that_s_24_12_1989 10.73084 49.40634  hopital
#> 8         2018_10_18 14.83382 48.39680    feild
#> 9               male 15.42775 48.32642 hospital
#> 10        01_12_2001 11.62814 46.54066 hospital
#> 11        2018_10_18 10.81086 47.20070    field
#> 12              <NA> 13.72975 48.47001    feild
#> 13              <NA> 10.98397 46.84982 hospital
#> 14 that_s_24_12_1989 14.10999 47.72840   clinic
#> 15              <NA> 13.87213 48.45742   clinic
#> 16        2018_10_17 13.21075 47.98304    feild
#> 17            female 12.49252 47.45841   clinic
#> 18        2018_10_17 11.04702 48.87581    feild
#> 19 that_s_24_12_1989 13.78844 48.74084    field
#> 20        2018_10_18 15.02981 47.83519    field

# add a wordlist
wordlist <- data.frame(
  from  = c("hopsital", "hopital",  "medical", "feild"),
  to    = c("hospital", "hospital", "clinic",  "field"),
  variable = rep("location", 4),
  stringsAsFactors = FALSE
)

clean_variables(toy_data,
                wordlists     = wordlist,
                spelling_vars = "variable"
               )

#>       'ID Date of Onset. DisCharge.. GENDER_  Épi.Case_définition
#> 1  dewyhn     2018-01-06  16_01_2018   female           confirmed
#> 2  cgkhox     2018-01-10  20_01_2018   female           confirmed
#> 3  qjvjib     2018-01-09  19_01_2018     male            probable
#> 4  oveifu     2018-01-04  14_01_2018     male          not_a_case
#> 5  aticgv     2018-01-03  13_01_2018   female           suspected
#> 6  pycbjk     2018-01-08  18_01_2018   female           confirmed
#> 7  sgapix     2018-01-02  12_01_2018   female           suspected
#> 8  fcnmip     2018-01-04  14_01_2018     male           confirmed
#> 9  fridsl     2018-01-07  17_01_2018     male           suspected
#> 10 bmqiyf     2018-01-02  12_01_2018   female          not_a_case
#> 11 vtptbs     2018-01-06  16_01_2018     male           suspected
#> 12 mhowzc     2018-01-02  12_01_2018   female          not_a_case
#> 13 ohoyig     2018-01-07  17_01_2018   female           suspected
#> 14 miylys     2018-01-10  20_01_2018   female          not_a_case
#> 15 bpktsa     2018-01-08  18_01_2018     male           confirmed
#> 16 pwojeh     2018-01-10  20_01_2018   female           confirmed
#> 17 hjaaws     2018-01-11  21_01_2018     male           suspected
#> 18 iwqrju     2018-01-09  19_01_2018   female            probable
#> 19 fulhnw     2018-01-03  13_01_2018     male            probable
#> 20 mbnyus     2018-01-09  19_01_2018   female          not_a_case
#>          messy/dates      lat      lon location
#> 1         2018_10_18 11.77449 48.63144 hospital
#> 2  that_s_24_12_1989 13.32132 47.51660 hospital
#> 3               male 11.67528 46.47367   clinic
#> 4  that_s_24_12_1989 12.33030 48.61684    field
#> 5         2018_10_18 14.24602 49.22929 hospital
#> 6             female 15.05639 48.15745    field
#> 7  that_s_24_12_1989 10.73084 49.40634 hospital
#> 8         2018_10_18 14.83382 48.39680    field
#> 9               male 15.42775 48.32642 hospital
#> 10        01_12_2001 11.62814 46.54066 hospital
#> 11        2018_10_18 10.81086 47.20070    field
#> 12              <NA> 13.72975 48.47001    field
#> 13              <NA> 10.98397 46.84982 hospital
#> 14 that_s_24_12_1989 14.10999 47.72840   clinic
#> 15              <NA> 13.87213 48.45742   clinic
#> 16        2018_10_17 13.21075 47.98304    field
#> 17            female 12.49252 47.45841   clinic
#> 18        2018_10_17 11.04702 48.87581    field
#> 19 that_s_24_12_1989 13.78844 48.74084    field
#> 20        2018_10_18 15.02981 47.83519    field