Clean variable labels and fix spelling according to a wordlist

Clean variable labels and fix spelling according to a wordlist

clean_variables(x, sep = "_", wordlists = NULL, spelling_vars = 3,
  sort_by = NULL, protect = FALSE, classes = NULL,
  warn_spelling = FALSE)

Arguments

x
a data.frame

sep
The separator used between words, and defaults to the underscore _.

wordlists
a data frame or named list of data frames with at least two columns defining the word list to be used. If this is a data frame, a third column must be present to split the wordlists by column in x (see spelling_vars).

spelling_vars
character or integer. If wordlists is a data frame, then this column in defines the columns in x corresponding to each section of the wordlists data frame. This defaults to 3, indicating the third column is to be used.

sort_by
a character the column to be used for sorting the values in each data frame. If the incoming variables are factors, this determines how the resulting factors will be sorted.

protect
a logical or numeric vector defining the columns to protect from any manipulation. Note: columns in protect will override any columns in either force_Date or guess_dates.

classes
a vector of class definitions for each of the columns. If this is not provided, the classes will be read from the columns themselves. Practically, this is used in clean_data() to mark columns as protected.

warn_spelling
if TRUE, errors and warnings from clean_spelling() will be aggregated and presented for each column that issues them. The default value is FALSE, which means that all errors and warnings will be ignored.

See also

Examples

## make toy data
toy_data <- messy_data(20)

# location data with mis-spellings, French, and English.
messy_locations <- c("hopsital", "h\u00f4pital", "hospital",
                     "m\u00e9dical", "clinic",
                     "feild", "field")
toy_data$location <- sample(messy_locations, 20, replace = TRUE)

## show data
toy_data

#>       'ID Date of Onset. DisCharge.. GENDER_  Épi.Case_définition
#> 1  rpxips     2018-01-07  17/01/2018   female           suspected
#> 2  vzjvks     2018-01-04  14/01/2018   female           suspected
#> 3  qayzyg     2018-01-07  17/01/2018   FEMALE           Confirmed
#> 4  rvtizy     2018-01-04  14/01/2018   female           Confirmed
#> 5  nkpdpq     2018-01-04  14/01/2018     male            probable
#> 6  efmgzr     2018-01-05  15/01/2018   female           confirmed
#> 7  opfjdk     2018-01-05  15/01/2018     male          Not.a.Case
#> 8  bqnkwh     2018-01-11  21/01/2018     Male           confirmed
#> 9  qidvtr     2018-01-07  17/01/2018     male          not a case
#> 10 fzrtqc     2018-01-10  20/01/2018   FEMALE           suspected
#> 11 boaeqv     2018-01-03  13/01/2018     Male           confirmed
#> 12 xvurim     2018-01-11  21/01/2018     MALE          not a case
#> 13 vuppem     2018-01-11  21/01/2018     Male         suspected  
#> 14 ybdloe     2018-01-09  19/01/2018     Male            probable
#> 15 ptbuxb     2018-01-10  20/01/2018     Male            PROBABLE
#> 16 sxteej     2018-01-04  14/01/2018     male         suspected  
#> 17 llehwg     2018-01-11  21/01/2018   FEMALE          Not.a.Case
#> 18 rwuems     2018-01-10  20/01/2018   FEMALE           confirmed
#> 19 hpjpkf     2018-01-06  16/01/2018     MALE          Not.a.Case
#> 20 tlcuca     2018-01-08  18/01/2018     Male            probable
#>           messy/dates       lat      lon location
#> 1                <NA> 12.634148 46.51891 hopsital
#> 2  that's 24/12/1989! 10.241484 48.72469    feild
#> 3          01-12-2001 15.501201 45.62249    feild
#> 4  that's 24/12/1989!  9.247526 47.93459    field
#> 5          2018_10_17 12.262425 47.53030 hospital
#> 6                <NA>  8.741405 47.71948  hôpital
#> 7                <NA> 13.675292 48.57484 hospital
#> 8              female  9.850850 47.79432  hôpital
#> 9     // 24//12//1989 12.607361 48.23488 hospital
#> 10               male 14.830120 48.11063 hospital
#> 11               male 15.157273 48.27420    field
#> 12               male 15.905606 49.57492  hôpital
#> 13         2018_10_17  9.693438 48.56777 hospital
#> 14    // 24//12//1989  7.476265 47.94039   clinic
#> 15             female 13.528738 47.85041    field
#> 16         2018-10-18 15.049143 48.26617   clinic
#> 17               male 11.681629 48.52840   clinic
#> 18         2018 10 19 12.227391 47.42061  hôpital
#> 19         2018 10 19 12.515002 48.94220 hopsital
#> 20               <NA> 13.441186 47.29364    feild

# clean labels
clean_variables(toy_data) # by default, it's the same as clean_variable_lables

#>       'ID Date of Onset. DisCharge.. GENDER_  Épi.Case_définition
#> 1  rpxips     2018-01-07  17_01_2018   female           suspected
#> 2  vzjvks     2018-01-04  14_01_2018   female           suspected
#> 3  qayzyg     2018-01-07  17_01_2018   female           confirmed
#> 4  rvtizy     2018-01-04  14_01_2018   female           confirmed
#> 5  nkpdpq     2018-01-04  14_01_2018     male            probable
#> 6  efmgzr     2018-01-05  15_01_2018   female           confirmed
#> 7  opfjdk     2018-01-05  15_01_2018     male          not_a_case
#> 8  bqnkwh     2018-01-11  21_01_2018     male           confirmed
#> 9  qidvtr     2018-01-07  17_01_2018     male          not_a_case
#> 10 fzrtqc     2018-01-10  20_01_2018   female           suspected
#> 11 boaeqv     2018-01-03  13_01_2018     male           confirmed
#> 12 xvurim     2018-01-11  21_01_2018     male          not_a_case
#> 13 vuppem     2018-01-11  21_01_2018     male           suspected
#> 14 ybdloe     2018-01-09  19_01_2018     male            probable
#> 15 ptbuxb     2018-01-10  20_01_2018     male            probable
#> 16 sxteej     2018-01-04  14_01_2018     male           suspected
#> 17 llehwg     2018-01-11  21_01_2018   female          not_a_case
#> 18 rwuems     2018-01-10  20_01_2018   female           confirmed
#> 19 hpjpkf     2018-01-06  16_01_2018     male          not_a_case
#> 20 tlcuca     2018-01-08  18_01_2018     male            probable
#>          messy/dates       lat      lon location
#> 1               <NA> 12.634148 46.51891 hopsital
#> 2  that_s_24_12_1989 10.241484 48.72469    feild
#> 3         01_12_2001 15.501201 45.62249    feild
#> 4  that_s_24_12_1989  9.247526 47.93459    field
#> 5         2018_10_17 12.262425 47.53030 hospital
#> 6               <NA>  8.741405 47.71948  hopital
#> 7               <NA> 13.675292 48.57484 hospital
#> 8             female  9.850850 47.79432  hopital
#> 9         24_12_1989 12.607361 48.23488 hospital
#> 10              male 14.830120 48.11063 hospital
#> 11              male 15.157273 48.27420    field
#> 12              male 15.905606 49.57492  hopital
#> 13        2018_10_17  9.693438 48.56777 hospital
#> 14        24_12_1989  7.476265 47.94039   clinic
#> 15            female 13.528738 47.85041    field
#> 16        2018_10_18 15.049143 48.26617   clinic
#> 17              male 11.681629 48.52840   clinic
#> 18        2018_10_19 12.227391 47.42061  hopital
#> 19        2018_10_19 12.515002 48.94220 hopsital
#> 20              <NA> 13.441186 47.29364    feild

# add a wordlist
wordlist <- data.frame(
  from  = c("hopsital", "hopital",  "medical", "feild"),
  to    = c("hospital", "hospital", "clinic",  "field"),
  variable = rep("location", 4),
  stringsAsFactors = FALSE
)

clean_variables(toy_data,
                wordlists     = wordlist,
                spelling_vars = "variable"
               )

#>       'ID Date of Onset. DisCharge.. GENDER_  Épi.Case_définition
#> 1  rpxips     2018-01-07  17_01_2018   female           suspected
#> 2  vzjvks     2018-01-04  14_01_2018   female           suspected
#> 3  qayzyg     2018-01-07  17_01_2018   female           confirmed
#> 4  rvtizy     2018-01-04  14_01_2018   female           confirmed
#> 5  nkpdpq     2018-01-04  14_01_2018     male            probable
#> 6  efmgzr     2018-01-05  15_01_2018   female           confirmed
#> 7  opfjdk     2018-01-05  15_01_2018     male          not_a_case
#> 8  bqnkwh     2018-01-11  21_01_2018     male           confirmed
#> 9  qidvtr     2018-01-07  17_01_2018     male          not_a_case
#> 10 fzrtqc     2018-01-10  20_01_2018   female           suspected
#> 11 boaeqv     2018-01-03  13_01_2018     male           confirmed
#> 12 xvurim     2018-01-11  21_01_2018     male          not_a_case
#> 13 vuppem     2018-01-11  21_01_2018     male           suspected
#> 14 ybdloe     2018-01-09  19_01_2018     male            probable
#> 15 ptbuxb     2018-01-10  20_01_2018     male            probable
#> 16 sxteej     2018-01-04  14_01_2018     male           suspected
#> 17 llehwg     2018-01-11  21_01_2018   female          not_a_case
#> 18 rwuems     2018-01-10  20_01_2018   female           confirmed
#> 19 hpjpkf     2018-01-06  16_01_2018     male          not_a_case
#> 20 tlcuca     2018-01-08  18_01_2018     male            probable
#>          messy/dates       lat      lon location
#> 1               <NA> 12.634148 46.51891 hospital
#> 2  that_s_24_12_1989 10.241484 48.72469    field
#> 3         01_12_2001 15.501201 45.62249    field
#> 4  that_s_24_12_1989  9.247526 47.93459    field
#> 5         2018_10_17 12.262425 47.53030 hospital
#> 6               <NA>  8.741405 47.71948 hospital
#> 7               <NA> 13.675292 48.57484 hospital
#> 8             female  9.850850 47.79432 hospital
#> 9         24_12_1989 12.607361 48.23488 hospital
#> 10              male 14.830120 48.11063 hospital
#> 11              male 15.157273 48.27420    field
#> 12              male 15.905606 49.57492 hospital
#> 13        2018_10_17  9.693438 48.56777 hospital
#> 14        24_12_1989  7.476265 47.94039   clinic
#> 15            female 13.528738 47.85041    field
#> 16        2018_10_18 15.049143 48.26617   clinic
#> 17              male 11.681629 48.52840   clinic
#> 18        2018_10_19 12.227391 47.42061 hospital
#> 19        2018_10_19 12.515002 48.94220 hospital
#> 20              <NA> 13.441186 47.29364    field