Try and guess dates from a characters

Note that THIS FEATURE IS STILL EXPERIMENTAL: we strongly recommend checking a few converted dates manually. This function tries to extract dates from a character vector or a factor. It treats each entry independently, using regular expressions to detect if a date is present, its format, and if successful it converts that entry to a standard Date with the Ymd format (e.g. 2018-01-21). Entries which cannot be processed result in NA. An error threshold can be used to define the maximum number of resulting NA (i.e. entries without an identified date) that can be tolerated. If this threshold is exceeded, the original vector is returned.

guess_dates(x, error_tolerance = 0.1, first_date = NULL,
  last_date = Sys.Date(), orders = list(world_named_months = c("Ybd",
  "dby"), world_digit_months = c("dmy", "Ymd"), US_formats = c("Omdy",
  "YOmd")), quiet = TRUE)

Arguments

x
a character vector or a factor

error_tolerance
a number between 0 and 1 indicating the proportion of entries which cannot be identified as dates to be tolerated; if this proportion is exceeded, the original vector is returned, and a message is issued; defaults to 0.1 (10 percent)

first_date
a Date object specifying the first valid date. Defaults to fifty years before the last_date.

last_date
a Date object specifying the last valid date. Defaults to the current date.

orders
date codes for fine-grained parsing of dates. This allows for parsing of mixed dates. If a list is supplied, that list will be used for successive tries in parsing. This is passed on to lubridate::parse_date_time(). Default orders parse World dmy/dby dates before US mdy/bdy dates.

quiet
a logical indicating if messages should be displayed to the console (TRUE, default); set to FALSE to silence messages

Details

Converting ambiguous character strings to dates is difficult for many reasons:

dates may not use the standard Ymd format

within the same variable, dates may follow different formats

dates may be mixed with things that are not dates

the behaviour of as.Date in the presence of non-date is hard to predict, sometimes returning NA, sometimes issuing an error.

This function tries to address all the above issues. Dates with the following format should be automatically detected, irrespective of separators (e.g. "-", " ", "/") and surrounding text:

"19 09 2018"

"2018 09 19"

"19 Sep 2018"

"2018 Sep 19"

Note that if a character string has multiple dates, it is currently hard to predict which date will be returned.

See also

Examples

x <- c("01-12-2001", "male", "female", "2018-10-18", NA, NA, "2018_10_17",
      "2018 10 19", "// 24/12/1989", "this is 24/12/1989!",
      "RECON NGO: 19 Sep 2018 :)", "6/9/11", "10/10/10")
FIRST_DATE <- as.Date("1969-11-11")
guess_dates(x, error_tolerance = 1, first_date = FIRST_DATE) # forced conversion

#>  [1] "2001-12-01" NA           NA           "2018-10-18" NA          
#>  [6] NA           "2018-10-17" "2018-10-19" "1989-12-24" "1989-12-24"
#> [11] "2018-09-19" "2011-09-06" "2010-10-10"

guess_dates(x, error_tolerance = 0.15, first_date = FIRST_DATE) # only 15% errors allowed

#>  [1] "01-12-2001"                "male"                     
#>  [3] "female"                    "2018-10-18"               
#>  [5] NA                          NA                         
#>  [7] "2018_10_17"                "2018 10 19"               
#>  [9] "// 24/12/1989"             "this is 24/12/1989!"      
#> [11] "RECON NGO: 19 Sep 2018 :)" "6/9/11"                   
#> [13] "10/10/10"