subset columns of a linelist object

This function works similarly to dplyr::select but can in addition refer to tagged variables. We recommend referring to variables and tags by their names: using logical and integer may have unexecpted results due to how regular variables and tagged variables are combined (see details). Functions in the form select_[tag] are wrappers for accessing specific tags as listed in tags_names().

# S3 method for linelist
select(.data, ..., warn = TRUE)

select_id(x)

select_date_onset(x)

select_date_reporting(x)

select_date_admission(x)

select_date_discharge(x)

select_date_outcome(x)

select_date_death(x)

select_gender(x)

select_age(x)

select_location(x)

select_occupation(x)

select_hcw(x)

select_outcome(x)

Arguments

.data

a linelist object

...

the variables to select, either using their column names, or tag names (or a mixture)

warn

a logical indicating if a warning should be issued if some tagged variables have been lost by the select operation; defaults to TRUE

x

a linelist object

Value

The function returns a linelist with selected columns.

Details

The function relies on a concatenation of two data.frames, the first one containing the regular variables, and the second one containing tagged variables. This approach is safe as long as regular variables and tagged variables are referred to by their names, which is the recommended approach. Other approaches using integer or logical may yield more confusing outputs, e.g. selecting all variables using TRUE will return the two data.frames concatenated.

See also

Author

Thibaut Jombart thibaut@data.org